refactor: MaybePendingCreator ergonomics + consolidate all tests into src/tests/
