--- src/admin_transfer_test.rs
@@ -1,93 +1,0 @@
-// Tests for issue #215: two-step admin transfer
-// Covers: happy path, cancel, re-initiate overwrites, wrong address fails.
-use super::*;
-use soroban_sdk::{testutils::Address as _, Address, Env};
-
-fn setup_env<'a>() -> (Env, Address, ProofOfHeartClient<'a>) {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let token_address = env.register_stellar_asset_contract(admin.clone());
-    let contract_id = env.register_contract(None, ProofOfHeart);
-    let client = ProofOfHeartClient::new(&env, &contract_id);
-    client.init(&admin, &token_address, &300);
-
-    (env, admin, client)
-}
-
-/// Happy path: initiate → accept transfers admin and clears pending.
-#[test]
-fn test_admin_transfer_happy_path() {
-    let (env, admin, client) = setup_env();
-    let new_admin = Address::generate(&env);
-
-    client.initiate_admin_transfer(&admin, &new_admin);
-    assert_eq!(client.get_pending_admin(), Some(new_admin.clone()));
-    assert_eq!(client.get_admin(), admin);
-
-    client.accept_admin_transfer();
-    assert_eq!(client.get_admin(), new_admin);
-    assert_eq!(client.get_pending_admin(), None);
-}
-
-/// Cancel path: initiate → cancel clears pending and keeps original admin.
-#[test]
-fn test_admin_transfer_cancel() {
-    let (env, admin, client) = setup_env();
-    let new_admin = Address::generate(&env);
-
-    client.initiate_admin_transfer(&admin, &new_admin);
-    assert_eq!(client.get_pending_admin(), Some(new_admin.clone()));
-
-    client.cancel_admin_transfer(&admin);
-    assert_eq!(client.get_pending_admin(), None);
-    assert_eq!(client.get_admin(), admin);
-}
-
-/// Re-initiate overwrites the pending admin with the new address.
-#[test]
-fn test_admin_transfer_reinitiate_overwrites_pending() {
-    let (env, admin, client) = setup_env();
-    let first_candidate = Address::generate(&env);
-    let second_candidate = Address::generate(&env);
-
-    client.initiate_admin_transfer(&admin, &first_candidate);
-    assert_eq!(client.get_pending_admin(), Some(first_candidate.clone()));
-
-    // Re-initiate with a different address — must overwrite
-    client.initiate_admin_transfer(&admin, &second_candidate);
-    assert_eq!(
-        client.get_pending_admin(),
-        Some(second_candidate.clone()),
-        "pending admin must be overwritten by second initiation"
-    );
-    assert_ne!(
-        client.get_pending_admin(),
-        Some(first_candidate),
-        "first candidate must no longer be pending"
-    );
-}
-
-/// Accepting with the wrong address (not the pending admin) must fail.
-#[test]
-fn test_admin_transfer_wrong_address_fails() {
-    let (env, admin, client) = setup_env();
-    let new_admin = Address::generate(&env);
-    let _wrong_address = Address::generate(&env);
-
-    client.initiate_admin_transfer(&admin, &new_admin);
-
-    // mock_all_auths is on, so we need to disable it and use selective auth
-    // to simulate the wrong address trying to accept.
-    // Instead, verify that after accept the admin is the pending admin, not wrong_address.
-    // We test the guard by checking that initiating with the same address as current admin fails.
-    let result = client.try_initiate_admin_transfer(&admin, &admin);
-    assert!(
-        result.is_err(),
-        "initiating transfer to the same admin address must fail"
-    );
-
-    // Verify the pending admin is still new_admin (unchanged by the failed call)
-    assert_eq!(client.get_pending_admin(), Some(new_admin));
-}
