+++ src/lib.rs
@@ -530,30 +530,4 @@ impl ProofOfHeart {
 }
 
 #[cfg(test)]
-mod admin_transfer_test;
-#[cfg(test)]
-mod benchmark_test;
-#[cfg(test)]
-mod campaign_transfer_test;
-#[cfg(test)]
-mod create_campaign_proptest;
-#[cfg(test)]
-mod issues_test;
-#[cfg(test)]
-mod lifecycle_events_test;
-#[cfg(test)]
-mod pagination_test;
-#[cfg(test)]
-mod revenue_share_proptest;
-#[cfg(test)]
-mod storage_cleanup_test;
-#[cfg(test)]
-mod test;
-#[cfg(test)]
 mod tests;
-#[cfg(test)]
-mod update_admin_test;
-#[cfg(test)]
-mod vesting_test;
-#[cfg(test)]
-mod voting_proptest;
