--- src/test.rs
@@ -1,4245 +1,0 @@
-use super::*;
-use soroban_sdk::token::Client as TokenClient;
-use soroban_sdk::token::StellarAssetClient as TokenAdminClient;
-use soroban_sdk::{
-    testutils::{Address as _, AuthorizedFunction, AuthorizedInvocation, Events, Ledger},
-    Address, Env, IntoVal, String, Symbol,
-};
-
-/// Test helper: build `CreateCampaignParams` with the same positional convention
-/// the old 9-argument `create_campaign` used, keeping test bodies readable.
-#[allow(clippy::too_many_arguments)]
-fn make_params(
-    creator: Address,
-    title: String,
-    description: String,
-    funding_goal: i128,
-    duration_days: u64,
-    category: Category,
-    has_revenue_sharing: bool,
-    revenue_share_percentage: u32,
-    max_contribution_per_user: i128,
-) -> CreateCampaignParams {
-    CreateCampaignParams {
-        creator,
-        title,
-        description,
-        funding_goal,
-        duration_days,
-        category,
-        has_revenue_sharing,
-        revenue_share_percentage,
-        max_contribution_per_user,
-    }
-}
-
-pub(crate) fn setup_env_with_default_min<'a>() -> (
-    Env,
-    Address,
-    Address,
-    Address,
-    Address,
-    TokenClient<'a>,
-    TokenAdminClient<'a>,
-    ProofOfHeartClient<'a>,
-) {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let creator = Address::generate(&env);
-    let contributor1 = Address::generate(&env);
-    let contributor2 = Address::generate(&env);
-
-    let token_address = env.register_stellar_asset_contract(admin.clone());
-    let token = TokenClient::new(&env, &token_address);
-    let token_admin = TokenAdminClient::new(&env, &token_address);
-
-    let contract_id = env.register_contract(None, ProofOfHeart);
-    let client = ProofOfHeartClient::new(&env, &contract_id);
-
-    client.init(&admin, &token_address, &300);
-
-    (
-        env,
-        admin,
-        creator,
-        contributor1,
-        contributor2,
-        token,
-        token_admin,
-        client,
-    )
-}
-
-pub(crate) fn setup_env<'a>() -> (
-    Env,
-    Address,
-    Address,
-    Address,
-    Address,
-    TokenClient<'a>,
-    TokenAdminClient<'a>,
-    ProofOfHeartClient<'a>,
-) {
-    let setup = setup_env_with_default_min();
-    setup.0.as_contract(&setup.7.address, || {
-        set_min_campaign_funding_goal(&setup.0, 1)
-    });
-    setup
-}
-
-#[test]
-fn test_init_only_once() {
-    let (_env, admin, _creator, _c1, _c2, token, _token_admin, client) = setup_env();
-    let res = client.try_init(&admin, &token.address, &300);
-    assert_eq!(res.unwrap_err().unwrap(), Error::AlreadyInitialized);
-}
-
-#[test]
-fn test_platform_fee_exact_storage() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let token_address = env.register_stellar_asset_contract(admin.clone());
-
-    let contract_id = env.register_contract(None, ProofOfHeart);
-    let client = ProofOfHeartClient::new(&env, &contract_id);
-
-    // Initialize with fee = 1000, should store exactly
-    client.init(&admin, &token_address, &1000);
-    assert_eq!(client.get_platform_fee(), 1000);
-}
-
-#[test]
-fn test_create_and_validation() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let title = String::from_str(&env, "Science Book");
-    let desc = String::from_str(&env, "Teaching science to kids");
-
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        0,
-        30,
-        Category::Publisher,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalMustBePositive);
-
-    // Test duration validation
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        500,
-        0,
-        Category::Publisher,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::InvalidDuration);
-
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        500,
-        400,
-        Category::Publisher,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::InvalidDuration);
-
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        500,
-        30,
-        Category::Educator,
-        true,
-        1000,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::RevenueShareOnlyForStartup);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        2000,
-        30,
-        Category::EducationalStartup,
-        true,
-        1500,
-        0i128,
-    ));
-    assert_eq!(campaign_id, 1);
-
-    let campaign = client.get_campaign(&campaign_id);
-    assert_eq!(campaign.id, 1);
-    assert_eq!(campaign.funding_goal, 2000);
-    assert!(campaign.is_active);
-    assert!(!campaign.is_verified);
-}
-
-#[test]
-fn test_min_campaign_funding_goal_boundary_and_admin_update() {
-    let (env, admin, creator, _c1, _c2, _token, _token_admin, client) =
-        setup_env_with_default_min();
-
-    assert_eq!(
-        client.get_min_campaign_funding_goal(),
-        CAMPAIGN_FUNDING_GOAL_MIN
-    );
-
-    let title = String::from_str(&env, "Minimum Goal");
-    let desc = String::from_str(&env, "Checks funding goal floor");
-
-    let below_min = CAMPAIGN_FUNDING_GOAL_MIN - 1;
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        below_min,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalTooLow);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        CAMPAIGN_FUNDING_GOAL_MIN,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(campaign_id, 1);
-
-    client.set_min_campaign_funding_goal(&admin, &500);
-    assert_eq!(client.get_min_campaign_funding_goal(), 500);
-
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        499,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalTooLow);
-}
-
-#[test]
-fn test_max_campaign_funding_goal_boundary_and_admin_update() {
-    let (env, admin, creator, _c1, _c2, _token, _token_admin, client) =
-        setup_env_with_default_min();
-
-    assert_eq!(
-        client.get_max_campaign_funding_goal(),
-        CAMPAIGN_FUNDING_GOAL_MAX
-    );
-
-    let title = String::from_str(&env, "Max Goal");
-    let desc = String::from_str(&env, "Checks funding goal ceiling");
-
-    // Exactly at the cap must succeed.
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        CAMPAIGN_FUNDING_GOAL_MAX,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(campaign_id, 1);
-
-    // One above the cap must fail.
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        CAMPAIGN_FUNDING_GOAL_MAX + 1,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalTooHigh);
-
-    // Admin raises the cap.
-    let new_max = CAMPAIGN_FUNDING_GOAL_MAX * 2;
-    client.set_max_campaign_funding_goal(&admin, &new_max);
-    assert_eq!(client.get_max_campaign_funding_goal(), new_max);
-
-    // Previously-rejected goal now succeeds.
-    let campaign_id2 = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        CAMPAIGN_FUNDING_GOAL_MAX + 1,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(campaign_id2, 2);
-}
-
-#[test]
-fn test_contribute_and_withdraw_success() {
-    let (env, admin, creator, contributor1, _, token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-
-    let title = String::from_str(&env, "Code Camp");
-    let desc = String::from_str(&env, "Learn Rust");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &1000);
-
-    assert_eq!(token.balance(&contributor1), 4000);
-    assert_eq!(token.balance(&client.address), 1000);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 1000);
-
-    client.withdraw_funds(&campaign_id);
-
-    assert_eq!(token.balance(&admin), 30);
-    assert_eq!(token.balance(&creator), 970);
-
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(!campaign.is_active);
-    assert!(campaign.funds_withdrawn);
-}
-
-#[test]
-fn test_creator_cannot_contribute_to_own_campaign() {
-    let (env, _admin, creator, _contributor1, _contributor2, _token, _token_admin, client) =
-        setup_env();
-
-    let title = String::from_str(&env, "Self Funding Block");
-    let desc = String::from_str(&env, "Creator should not contribute");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    let res = client.try_contribute(&campaign_id, &creator, &100);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NotAuthorized);
-}
-
-#[test]
-fn test_cancel_and_refund() {
-    let (env, _admin, creator, contributor1, contributor2, token, token_admin, client) =
-        setup_env();
-
-    token_admin.mint(&contributor1, &2000);
-    token_admin.mint(&contributor2, &1000);
-
-    let title = String::from_str(&env, "Failed Idea");
-    let desc = String::from_str(&env, "Desc");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        5000,
-        10,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.contribute(&campaign_id, &contributor2, &500);
-
-    client.cancel_campaign(&campaign_id);
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(campaign.is_cancelled);
-
-    client.claim_refund(&campaign_id, &contributor1);
-    client.claim_refund(&campaign_id, &contributor2);
-
-    assert_eq!(token.balance(&contributor1), 2000);
-    assert_eq!(token.balance(&contributor2), 1000);
-    assert_eq!(token.balance(&client.address), 0);
-}
-
-#[test]
-fn test_claim_refund_requires_contributor_auth() {
-    let (env, _admin, creator, contributor1, _contributor2, token, token_admin, client) =
-        setup_env();
-
-    token_admin.mint(&contributor1, &2000);
-
-    let title = String::from_str(&env, "Auth Refund");
-    let desc = String::from_str(&env, "Only contributor can claim");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        5000,
-        10,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.cancel_campaign(&campaign_id);
-
-    client.claim_refund(&campaign_id, &contributor1);
-
-    let auths = env.auths();
-    assert_eq!(auths.len(), 1);
-    let (auth_addr, invocation) = &auths[0];
-    assert_eq!(auth_addr, &contributor1);
-    assert_eq!(
-        invocation,
-        &AuthorizedInvocation {
-            function: AuthorizedFunction::Contract((
-                client.address.clone(),
-                Symbol::new(&env, "claim_refund"),
-                (campaign_id, contributor1.clone()).into_val(&env),
-            )),
-            sub_invocations: Default::default(),
-        }
-    );
-
-    assert_eq!(token.balance(&contributor1), 2000);
-}
-
-#[test]
-fn test_pull_based_revenue_distribution() {
-    let (env, _admin, creator, contributor1, contributor2, token, token_admin, client) =
-        setup_env();
-
-    token_admin.mint(&contributor1, &1000);
-    token_admin.mint(&contributor2, &1000);
-    token_admin.mint(&creator, &10000);
-
-    let title = String::from_str(&env, "Next Gen AI");
-    let desc = String::from_str(&env, "Build AI");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        2000,
-        30,
-        Category::EducationalStartup,
-        true,
-        2000,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.contribute(&campaign_id, &contributor2, &1000);
-
-    client.withdraw_funds(&campaign_id);
-
-    // Deposit revenue
-    token_admin.mint(&creator, &5000);
-    client.deposit_revenue(&campaign_id, &5000);
-    assert_eq!(client.get_revenue_pool(&campaign_id), 5000);
-
-    // contributor_pool = (5000 * 2000) / 10000 = 1000 (20% of pool to contributors)
-    // contributor1 share = (1000 * 1000) / 2000 = 500
-    client.claim_revenue(&campaign_id, &contributor1);
-    assert_eq!(token.balance(&contributor1), 500);
-    assert_eq!(client.get_revenue_claimed(&campaign_id, &contributor1), 500);
-
-    client.deposit_revenue(&campaign_id, &1000);
-    assert_eq!(client.get_revenue_pool(&campaign_id), 6000);
-
-    // contributor1: total_due = (1000 * 1200) / 2000 = 600, already_claimed = 500, claimable = 100
-    client.claim_revenue(&campaign_id, &contributor1);
-    assert_eq!(token.balance(&contributor1), 600);
-
-    // contributor2: total_due = (1000 * 1200) / 2000 = 600, claimable = 600
-    client.claim_revenue(&campaign_id, &contributor2);
-    assert_eq!(token.balance(&contributor2), 600);
-}
-
-#[test]
-fn test_failure_states() {
-    let (env, _admin, creator, contributor1, _, token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5000);
-
-    let title = String::from_str(&env, "Deadline Test");
-    let desc = String::from_str(&env, "Desc");
-    let duration_days = 2;
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        duration_days,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    let res = client.try_withdraw_funds(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NoFundsToWithdraw);
-
-    client.contribute(&campaign_id, &contributor1, &500);
-
-    let res = client.try_withdraw_funds(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalNotReached);
-
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: env.ledger().timestamp() + (duration_days * 86450),
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence(),
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    let res = client.try_contribute(&campaign_id, &contributor1, &500);
-    assert_eq!(res.unwrap_err().unwrap(), Error::DeadlinePassed);
-
-    let res = client.try_withdraw_funds(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalNotReached);
-
-    // After failure refund successful
-    client.claim_refund(&campaign_id, &contributor1);
-    assert_eq!(token.balance(&contributor1), 5000);
-}
-
-#[test]
-fn test_multiple_concurrent_campaigns_are_isolated() {
-    let (env, _admin, creator1, contributor1, contributor2, token, token_admin, client) =
-        setup_env();
-
-    let creator2 = Address::generate(&env);
-    let creator3 = Address::generate(&env);
-
-    token_admin.mint(&contributor1, &10000);
-    token_admin.mint(&contributor2, &10000);
-    token_admin.mint(&creator3, &10000);
-
-    let c1_title = String::from_str(&env, "Campaign 1");
-    let c1_desc = String::from_str(&env, "Educator campaign");
-    let campaign_1 = client.create_campaign(&make_params(
-        creator1.clone(),
-        c1_title.clone(),
-        c1_desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_1);
-
-    let c2_title = String::from_str(&env, "Campaign 2");
-    let c2_desc = String::from_str(&env, "Learner campaign");
-    let campaign_2 = client.create_campaign(&make_params(
-        creator2.clone(),
-        c2_title.clone(),
-        c2_desc.clone(),
-        1500,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_2);
-
-    let c3_title = String::from_str(&env, "Campaign 3");
-    let c3_desc = String::from_str(&env, "Startup campaign");
-    let campaign_3 = client.create_campaign(&make_params(
-        creator3.clone(),
-        c3_title.clone(),
-        c3_desc.clone(),
-        2000,
-        30,
-        Category::EducationalStartup,
-        true,
-        1500,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_3);
-
-    assert_eq!(campaign_1, 1);
-    assert_eq!(campaign_2, 2);
-    assert_eq!(campaign_3, 3);
-    assert_eq!(client.get_campaign_count(), 3);
-
-    client.contribute(&campaign_1, &contributor1, &1000);
-
-    client.contribute(&campaign_2, &contributor1, &400);
-    client.contribute(&campaign_2, &contributor2, &500);
-
-    client.contribute(&campaign_3, &contributor1, &1200);
-    client.contribute(&campaign_3, &contributor2, &800);
-
-    assert_eq!(client.get_contribution(&campaign_1, &contributor1), 1000);
-    assert_eq!(client.get_contribution(&campaign_1, &contributor2), 0);
-    assert_eq!(client.get_contribution(&campaign_2, &contributor1), 400);
-    assert_eq!(client.get_contribution(&campaign_2, &contributor2), 500);
-    assert_eq!(client.get_contribution(&campaign_3, &contributor1), 1200);
-    assert_eq!(client.get_contribution(&campaign_3, &contributor2), 800);
-
-    client.withdraw_funds(&campaign_1);
-
-    let c1_after_withdraw = client.get_campaign(&campaign_1);
-    let c2_after_withdraw = client.get_campaign(&campaign_2);
-    let c3_after_withdraw = client.get_campaign(&campaign_3);
-
-    assert!(c1_after_withdraw.funds_withdrawn);
-    assert!(!c1_after_withdraw.is_active);
-
-    assert_eq!(c2_after_withdraw.amount_raised, 900);
-    assert!(!c2_after_withdraw.funds_withdrawn);
-    assert!(c2_after_withdraw.is_active);
-    assert!(!c2_after_withdraw.is_cancelled);
-
-    assert_eq!(c3_after_withdraw.amount_raised, 2000);
-    assert!(!c3_after_withdraw.funds_withdrawn);
-    assert!(c3_after_withdraw.is_active);
-    assert!(!c3_after_withdraw.is_cancelled);
-
-    client.cancel_campaign(&campaign_2);
-
-    let c1_after_cancel = client.get_campaign(&campaign_1);
-    let c2_after_cancel = client.get_campaign(&campaign_2);
-    let c3_after_cancel = client.get_campaign(&campaign_3);
-
-    assert!(c2_after_cancel.is_cancelled);
-    assert!(!c2_after_cancel.is_active);
-
-    assert!(c1_after_cancel.funds_withdrawn);
-    assert!(!c1_after_cancel.is_cancelled);
-    assert!(c3_after_cancel.is_active);
-    assert!(!c3_after_cancel.is_cancelled);
-
-    assert_eq!(client.get_revenue_pool(&campaign_1), 0);
-    assert_eq!(client.get_revenue_pool(&campaign_2), 0);
-
-    client.withdraw_funds(&campaign_3);
-    let c3_after_withdraw_funds = client.get_campaign(&campaign_3);
-    assert!(c3_after_withdraw_funds.funds_withdrawn);
-    assert!(!c3_after_withdraw_funds.is_active);
-
-    client.deposit_revenue(&campaign_3, &3000);
-
-    assert_eq!(client.get_revenue_pool(&campaign_1), 0);
-    assert_eq!(client.get_revenue_pool(&campaign_2), 0);
-    assert_eq!(client.get_revenue_pool(&campaign_3), 3000);
-
-    // Balance checks to ensure campaign operations remained isolated
-    assert_eq!(token.balance(&client.address), 3900);
-    assert_eq!(token.balance(&creator3), 8940);
-}
-
-#[test]
-fn test_double_refund_prevention() {
-    let (env, _admin, creator, contributor1, _, token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &2000);
-
-    let title = String::from_str(&env, "Double Refund");
-    let desc = String::from_str(&env, "Test double refund");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        5000,
-        10,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.cancel_campaign(&campaign_id);
-
-    client.claim_refund(&campaign_id, &contributor1);
-    assert_eq!(token.balance(&contributor1), 2000);
-
-    let res = client.try_claim_refund(&campaign_id, &contributor1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NoFundsToWithdraw);
-    assert_eq!(token.balance(&contributor1), 2000);
-}
-
-#[test]
-fn test_get_version() {
-    let (_env, _admin, _creator, _contributor1, _contributor2, _token, _token_admin, client) =
-        setup_env();
-
-    // init stores CONTRACT_VERSION (1) in instance storage
-    assert_eq!(client.get_version(), 1u32);
-}
-
-#[test]
-fn test_admin_verify_campaign_success() {
-    let (env, _admin, creator, _contributor1, _contributor2, _token, _token_admin, client) =
-        setup_env();
-
-    let title = String::from_str(&env, "Admin Verification");
-    let desc = String::from_str(&env, "Admin verifies campaign");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.verify_campaign(&campaign_id);
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(campaign.is_verified);
-}
-
-#[test]
-fn test_update_campaign_blocks_after_admin_verification() {
-    let (env, _admin, creator, _contributor1, _contributor2, _token, _token_admin, client) =
-        setup_env();
-
-    let title = String::from_str(&env, "Original Title");
-    let desc = String::from_str(&env, "Original Description");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.verify_campaign(&campaign_id);
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(campaign.is_verified);
-
-    // Fix #416: update_campaign must be blocked after admin verification.
-    let new_title = String::from_str(&env, "New Title");
-    let new_desc = String::from_str(&env, "New Description");
-    let res = client.try_update_campaign(&campaign_id, &new_title, &new_desc);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignAlreadyVerified);
-
-    // Verify the campaign content is unchanged.
-    let campaign_after = client.get_campaign(&campaign_id);
-    assert_eq!(campaign_after.title, title);
-    assert_eq!(campaign_after.description, desc);
-}
-
-#[test]
-fn test_update_campaign_blocks_after_community_verification() {
-    let (env, _admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-    let voter3 = Address::generate(&env);
-
-    token_admin.mint(&contributor1, &100);
-    token_admin.mint(&contributor2, &100);
-    token_admin.mint(&voter3, &100);
-
-    let title = String::from_str(&env, "Original Title");
-    let desc = String::from_str(&env, "Original Description");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-    client.vote_on_campaign(&campaign_id, &contributor2, &true);
-    client.vote_on_campaign(&campaign_id, &voter3, &true);
-
-    client.verify_campaign_with_votes(&campaign_id);
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(campaign.is_verified);
-
-    // Fix #416: update_campaign must be blocked after community verification.
-    let new_title = String::from_str(&env, "New Title");
-    let new_desc = String::from_str(&env, "New Description");
-    let res = client.try_update_campaign(&campaign_id, &new_title, &new_desc);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignAlreadyVerified);
-
-    let campaign_after = client.get_campaign(&campaign_id);
-    assert_eq!(campaign_after.title, title);
-    assert_eq!(campaign_after.description, desc);
-}
-
-#[test]
-fn test_admin_verify_campaign_duplicate_attempt() {
-    let (env, _admin, creator, _contributor1, _contributor2, _token, _token_admin, client) =
-        setup_env();
-
-    let title = String::from_str(&env, "Duplicate Verification");
-    let desc = String::from_str(&env, "Cannot verify twice");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Publisher,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.verify_campaign(&campaign_id);
-    let res = client.try_verify_campaign(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::AdminVerificationConflict);
-}
-
-#[test]
-fn test_community_voting_verification_success() {
-    let (env, _admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-    let voter3 = Address::generate(&env);
-
-    token_admin.mint(&contributor1, &100);
-    token_admin.mint(&contributor2, &100);
-    token_admin.mint(&voter3, &100);
-
-    let title = String::from_str(&env, "Community Verified");
-    let desc = String::from_str(&env, "Verify by voting");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-    client.vote_on_campaign(&campaign_id, &contributor2, &true);
-    client.vote_on_campaign(&campaign_id, &voter3, &false);
-
-    assert_eq!(client.get_approve_votes(&campaign_id), 2);
-    assert_eq!(client.get_reject_votes(&campaign_id), 1);
-    assert!(client.has_voted(&campaign_id, &contributor1));
-
-    client.verify_campaign_with_votes(&campaign_id);
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(campaign.is_verified);
-
-    let res = client.try_verify_campaign_with_votes(&campaign_id);
-    assert_eq!(
-        res.unwrap_err().unwrap(),
-        Error::CommunityVerificationConflict
-    );
-}
-
-#[test]
-fn test_vote_prevents_double_voting_and_requires_token_holder() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    let non_holder = Address::generate(&env);
-
-    token_admin.mint(&contributor1, &100);
-
-    let title = String::from_str(&env, "Vote Safety");
-    let desc = String::from_str(&env, "No duplicate votes");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        500,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &false);
-    assert_eq!(res.unwrap_err().unwrap(), Error::AlreadyVoted);
-
-    let res = client.try_vote_on_campaign(&campaign_id, &non_holder, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NotTokenHolder);
-}
-
-#[test]
-fn test_verify_campaign_quorum_and_threshold_edges() {
-    let (env, admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-    let voter3 = Address::generate(&env);
-    let voter4 = Address::generate(&env);
-
-    token_admin.mint(&contributor1, &100);
-    token_admin.mint(&contributor2, &100);
-    token_admin.mint(&voter3, &100);
-    token_admin.mint(&voter4, &100);
-
-    client.set_voting_params(&admin, &4, &7500);
-    assert_eq!(client.get_min_votes_quorum(), 4);
-    assert_eq!(client.get_approval_threshold_bps(), 7500);
-
-    let title1 = String::from_str(&env, "Quorum Campaign");
-    let desc1 = String::from_str(&env, "Needs 4 votes");
-    let campaign_id_1 = client.create_campaign(&make_params(
-        creator.clone(),
-        title1.clone(),
-        desc1.clone(),
-        700,
-        30,
-        Category::Publisher,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.vote_on_campaign(&campaign_id_1, &contributor1, &true);
-    client.vote_on_campaign(&campaign_id_1, &contributor2, &true);
-    client.vote_on_campaign(&campaign_id_1, &voter3, &true);
-
-    let res = client.try_verify_campaign_with_votes(&campaign_id_1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::VotingQuorumNotMet);
-
-    client.vote_on_campaign(&campaign_id_1, &voter4, &false);
-    client.verify_campaign(&campaign_id_1);
-    assert!(client.get_campaign(&campaign_id_1).is_verified);
-
-    let title2 = String::from_str(&env, "Threshold Campaign");
-    let desc2 = String::from_str(&env, "Fails threshold");
-    let campaign_id_2 = client.create_campaign(&make_params(
-        creator.clone(),
-        title2.clone(),
-        desc2.clone(),
-        700,
-        30,
-        Category::Publisher,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.vote_on_campaign(&campaign_id_2, &contributor1, &true);
-    client.vote_on_campaign(&campaign_id_2, &contributor2, &true);
-    client.vote_on_campaign(&campaign_id_2, &voter3, &false);
-    client.vote_on_campaign(&campaign_id_2, &voter4, &false);
-
-    let res = client.try_verify_campaign_with_votes(&campaign_id_2);
-    assert_eq!(res.unwrap_err().unwrap(), Error::VotingThresholdNotMet);
-}
-
-#[test]
-fn test_update_platform_fee() {
-    let (env, _admin, _creator, _contributor1, _contributor2, _token, _token_admin, client) =
-        setup_env();
-
-    let result = client.try_update_platform_fee(&500);
-    assert!(
-        result.is_ok(),
-        "Admin should be able to update platform fee"
-    );
-
-    let events = env.events().all();
-    let last_event = events.last().unwrap();
-    let expected_topics = (String::from_str(&env, "fee_updated"),).into_val(&env);
-    assert_eq!(last_event.1, expected_topics);
-
-    let data_vec: soroban_sdk::Vec<u32> = soroban_sdk::FromVal::from_val(&env, &last_event.2);
-    assert_eq!(data_vec.get(0).unwrap(), 300);
-    assert_eq!(data_vec.get(1).unwrap(), 500);
-
-    // Issue #343 / #401: fees above the cap are rejected with InvalidPlatformFee.
-    // (The earlier duplicate ValidationFailed assertion was unreachable because
-    // update_platform_fee only ever returned InvalidPlatformFee — see #401.)
-    let result = client.try_update_platform_fee(&5000);
-    assert_eq!(result.unwrap_err().unwrap(), Error::InvalidPlatformFee);
-}
-
-#[test]
-fn test_get_campaign_not_found() {
-    let (_env, _admin, _creator, _contributor1, _contributor2, _token, _token_admin, client) =
-        setup_env();
-
-    let res = client.try_get_campaign(&999);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotFound);
-}
-
-#[test]
-fn test_deadline_boundary() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5000);
-
-    let title = String::from_str(&env, "Boundary Test");
-    let desc = String::from_str(&env, "Testing exact deadline boundary");
-    let duration_days = 2;
-    let funding_goal = 1000;
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        funding_goal,
-        duration_days,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    let campaign = client.get_campaign(&campaign_id);
-    let deadline = campaign.deadline;
-
-    // Fast forward to exactly the deadline
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: deadline,
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence(),
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    // Should succeed exactly at the deadline
-    client.contribute(&campaign_id, &contributor1, &500);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 500);
-
-    // Fast forward to exactly 1 second past the deadline
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: deadline + 1,
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence(),
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    // Should fail past the deadline
-    let res = client.try_contribute(&campaign_id, &contributor1, &500);
-    assert_eq!(res.unwrap_err().unwrap(), Error::DeadlinePassed);
-}
-
-#[test]
-fn test_reinit_prevention() {
-    let (env, admin, _, _, _, token, _, client) = setup_env();
-
-    let attacker = Address::generate(&env);
-    let fake_token = Address::generate(&env);
-
-    // Attempt re-initialization with different values — must be rejected
-    let res = client.try_init(&attacker, &fake_token, &0);
-    assert!(res.is_err()); // Should fail with AlreadyInitialized
-
-    // Verify original values remain unchanged after rejected re-init
-    assert_eq!(client.get_admin(), admin);
-    assert_eq!(client.get_token(), token.address);
-    assert_eq!(client.get_platform_fee(), 300);
-}
-
-#[test]
-fn test_initialization_getters() {
-    let (_, admin, _, _, _, token, _, client) = setup_env();
-
-    assert_eq!(client.get_admin(), admin);
-    assert_eq!(client.get_token(), token.address);
-    assert_eq!(client.get_platform_fee(), 300);
-    assert_eq!(client.get_campaign_count(), 0);
-}
-
-#[test]
-fn test_revenue_sharing_edge_cases() {
-    let (env, _admin, creator, contributor1, contributor2, token, token_admin, client) =
-        setup_env();
-
-    // 1. Non-revenue campaign: check ValidationFailed
-    let title_nr = String::from_str(&env, "No Revenue");
-    let desc_nr = String::from_str(&env, "Non-revenue campaign");
-    let campaign_nr = client.create_campaign(&make_params(
-        creator.clone(),
-        title_nr.clone(),
-        desc_nr.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_nr);
-    let res = client.try_claim_revenue(&campaign_nr, &contributor1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-
-    token_admin.mint(&contributor1, &10);
-    token_admin.mint(&contributor2, &10);
-    token_admin.mint(&creator, &100);
-
-    let title = String::from_str(&env, "Rounding Test");
-    let desc = String::from_str(&env, "Test rounding and pool edge cases");
-    // 50% revenue share to contributors (5000 bps = max allowed)
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        3,
-        30,
-        Category::EducationalStartup,
-        true,
-        5000,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &1);
-    client.contribute(&campaign_id, &contributor2, &2);
-    client.withdraw_funds(&campaign_id);
-
-    // 2. Zero pool: should fail NoFundsToWithdraw
-    let res = client.try_claim_revenue(&campaign_id, &contributor1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NoFundsToWithdraw);
-
-    // 3. Rounding: pool=10, contributor_pool = (10*5000)/10000 = 5
-    // contributor1 (1 of 3): (1 * 5) / 3 = 1
-    // contributor2 (2 of 3): (2 * 5) / 3 = 3
-    client.deposit_revenue(&campaign_id, &10);
-    client.claim_revenue(&campaign_id, &contributor1);
-    assert_eq!(token.balance(&contributor1), 10); // Initial 10 - 1 contribution + 1 claimed
-
-    client.claim_revenue(&campaign_id, &contributor2);
-    assert_eq!(token.balance(&contributor2), 11); // Initial 10 - 2 contribution + 3 claimed
-
-    // 4. Double claim: NoFundsToWithdraw
-    let res = client.try_claim_revenue(&campaign_id, &contributor1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NoFundsToWithdraw);
-}
-
-// ── Issue #101 ────────────────────────────────────────────────────────────────
-// Regression: campaign_count must never reset after deployment, regardless of
-// which admin-level operations are executed.
-#[test]
-fn test_campaign_count_cannot_reset_after_deployment() {
-    let (env, _admin, creator, _, _, token, _, client) = setup_env();
-
-    // Start at zero
-    assert_eq!(client.get_campaign_count(), 0);
-
-    // Create three campaigns so the counter increments
-    for i in 1u32..=3 {
-        let title = String::from_str(&env, "Campaign");
-        let desc = String::from_str(&env, "Description");
-        let id = client.create_campaign(&make_params(
-            creator.clone(),
-            title.clone(),
-            desc.clone(),
-            1000,
-            30,
-            Category::Educator,
-            false,
-            0,
-            0i128,
-        ));
-        assert_eq!(id, i);
-    }
-    assert_eq!(client.get_campaign_count(), 3);
-
-    // Admin flows that must NOT reset the counter
-    client.update_platform_fee(&500);
-    assert_eq!(client.get_campaign_count(), 3);
-
-    let new_admin = Address::generate(&env);
-    client.update_admin(&new_admin);
-    client.accept_admin_transfer();
-    assert_eq!(client.get_campaign_count(), 3);
-
-    // Re-initialisation attempt must be rejected and counter must stay intact
-    let res = client.try_init(&new_admin, &token.address, &300);
-    assert_eq!(res.unwrap_err().unwrap(), Error::AlreadyInitialized);
-    assert_eq!(client.get_campaign_count(), 3);
-}
-
-// ── Issue #112 ────────────────────────────────────────────────────────────────
-// Negative test: approval_threshold_bps values above 10 000 must be rejected.
-#[test]
-fn test_set_voting_params_rejects_threshold_over_10000() {
-    let (_env, admin, _, _, _, _, _, client) = setup_env();
-
-    // Exactly 10 000 is the boundary — must be accepted
-    let res = client.try_set_voting_params(&admin, &3, &10000);
-    assert!(
-        res.is_ok(),
-        "approval_threshold_bps = 10000 should be valid"
-    );
-
-    // 10 001 exceeds the maximum — must be rejected
-    let res = client.try_set_voting_params(&admin, &3, &10001);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-
-    // u32::MAX is far above the maximum — must be rejected
-    let res = client.try_set_voting_params(&admin, &3, &u32::MAX);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-}
-
-// ── Issue #129 ────────────────────────────────────────────────────────────────
-// Invariant: the sum of every individual contributor's stored contribution must
-// equal the campaign's amount_raised at all times.
-#[test]
-fn test_contribution_accounting_invariant() {
-    let (env, _admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-
-    let contributor3 = Address::generate(&env);
-
-    token_admin.mint(&contributor1, &3000);
-    token_admin.mint(&contributor2, &3000);
-    token_admin.mint(&contributor3, &3000);
-
-    let title = String::from_str(&env, "Invariant Campaign");
-    let desc = String::from_str(&env, "Accounting invariant check");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        5000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    // Each contributor makes two separate contributions
-    client.contribute(&campaign_id, &contributor1, &500);
-    client.contribute(&campaign_id, &contributor2, &750);
-    client.contribute(&campaign_id, &contributor3, &250);
-    client.contribute(&campaign_id, &contributor1, &300);
-    client.contribute(&campaign_id, &contributor2, &200);
-
-    let c1 = client.get_contribution(&campaign_id, &contributor1);
-    let c2 = client.get_contribution(&campaign_id, &contributor2);
-    let c3 = client.get_contribution(&campaign_id, &contributor3);
-
-    assert_eq!(c1, 800, "contributor1 total must be 800");
-    assert_eq!(c2, 950, "contributor2 total must be 950");
-    assert_eq!(c3, 250, "contributor3 total must be 250");
-
-    // The canonical invariant: per-user totals must sum to amount_raised
-    let campaign = client.get_campaign(&campaign_id);
-    assert_eq!(
-        c1 + c2 + c3,
-        campaign.amount_raised,
-        "sum of per-user contributions must equal campaign amount_raised"
-    );
-}
-
-// ── Issue #117 ────────────────────────────────────────────────────────────────
-// Auth test: a non-admin caller must not be able to update voting parameters.
-#[test]
-fn test_set_voting_params_rejects_non_admin() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    // creator is not the admin — the call must be rejected
-    let res = client.try_set_voting_params(&creator, &5, &7000);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NotAuthorized);
-
-    // A completely random address also must be rejected
-    let random = Address::generate(&env);
-    let res = client.try_set_voting_params(&random, &5, &7000);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NotAuthorized);
-}
-
-#[test]
-fn test_view_functions_error_handling() {
-    let (env, _admin, creator, contributor1, _, _, _, client) = setup_env();
-
-    // Create a valid campaign for relative testing
-    let title = String::from_str(&env, "View Test");
-    let desc = String::from_str(&env, "Testing view functions");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    let stranger = Address::generate(&env);
-    let invalid_id = 999u32;
-
-    // 1. get_campaign with invalid ID
-    // Expected: Returns Error::CampaignNotFound (previously panicked before Issue #8 fix)
-    let res = client.try_get_campaign(&invalid_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotFound);
-
-    // 2. get_contribution with valid campaign but non-existent contributor
-    // Expected: Returns 0 (no panic)
-    assert_eq!(client.get_contribution(&campaign_id, &stranger), 0);
-
-    // 3. get_contribution with invalid campaign ID
-    // Expected: Returns 0 (no panic)
-    assert_eq!(client.get_contribution(&invalid_id, &contributor1), 0);
-
-    // 4. get_revenue_pool with invalid campaign ID
-    // Expected: Returns 0 (no panic)
-    assert_eq!(client.get_revenue_pool(&invalid_id), 0);
-
-    // 5. get_revenue_claimed with valid campaign but non-existent contributor
-    // Expected: Returns 0 (no panic)
-    assert_eq!(client.get_revenue_claimed(&campaign_id, &stranger), 0);
-
-    // 6. get_revenue_claimed with invalid campaign ID
-    // Expected: Returns 0 (no panic)
-    assert_eq!(client.get_revenue_claimed(&invalid_id, &contributor1), 0);
-}
-
-// ── Issue #68: update_campaign_description ────────────────────────────────────
-
-#[test]
-fn test_update_campaign_description_success() {
-    let (env, _admin, creator, _contributor1, _, _token, _token_admin, client) = setup_env();
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Original Title").clone(),
-        String::from_str(&env, "Original description").clone(),
-        1_000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    let new_desc = String::from_str(&env, "Updated description with more detail");
-    let res = client.try_update_campaign_description(&campaign_id, &new_desc);
-    assert!(res.is_ok());
-
-    let campaign = client.get_campaign(&campaign_id);
-    assert_eq!(campaign.description, new_desc);
-    // Funding goal and deadline must remain unchanged
-    assert_eq!(campaign.funding_goal, 1_000);
-}
-
-#[test]
-fn test_update_campaign_description_rejects_cancelled() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Title").clone(),
-        String::from_str(&env, "Desc").clone(),
-        1_000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.cancel_campaign(&campaign_id);
-
-    let res =
-        client.try_update_campaign_description(&campaign_id, &String::from_str(&env, "New desc"));
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotActive);
-}
-
-// ── Issue #99: init idempotency regression tests ──────────────────────────────
-
-#[test]
-fn test_init_returns_already_initialized_error() {
-    let (_env, admin, _creator, _c1, _c2, token, _token_admin, client) = setup_env();
-    let err = client
-        .try_init(&admin, &token.address, &300)
-        .unwrap_err()
-        .unwrap();
-    assert_eq!(err, Error::AlreadyInitialized);
-}
-
-#[test]
-fn test_init_preserves_all_config_state() {
-    let (_env, admin, _creator, _c1, _c2, token, _token_admin, client) = setup_env();
-
-    // Attempt re-init with completely different parameters
-    let _ = client.try_init(&admin, &token.address, &999);
-
-    // Every stored value must reflect the original init, not the rejected one
-    assert_eq!(client.get_admin(), admin);
-    assert_eq!(client.get_token(), token.address);
-    assert_eq!(client.get_platform_fee(), 300);
-    assert_eq!(client.get_campaign_count(), 0);
-    assert_eq!(client.get_version(), 1);
-    assert_eq!(
-        client.get_min_votes_quorum(),
-        crate::voting::DEFAULT_MIN_VOTES_QUORUM
-    );
-    assert_eq!(
-        client.get_approval_threshold_bps(),
-        crate::voting::DEFAULT_APPROVAL_THRESHOLD_BPS
-    );
-}
-
-#[test]
-fn test_init_rejects_every_subsequent_call() {
-    let (_env, admin, _creator, _c1, _c2, token, _token_admin, client) = setup_env();
-
-    // Each extra call must be rejected regardless of how many times it is attempted
-    for _ in 0..3 {
-        let res = client.try_init(&admin, &token.address, &300);
-        assert_eq!(
-            res.unwrap_err().unwrap(),
-            Error::AlreadyInitialized,
-            "expected AlreadyInitialized on every repeated call"
-        );
-    }
-}
-
-#[test]
-fn test_init_cannot_overwrite_after_campaign_created() {
-    let (env, admin, creator, _c1, _c2, token, _token_admin, client) = setup_env();
-
-    // Advance state: create a campaign so campaign_count > 0
-    let _ = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Test Campaign").clone(),
-        String::from_str(&env, "Testing init idempotency after state change").clone(),
-        1_000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(client.get_campaign_count(), 1);
-
-    // Attempt re-init must still be rejected
-    let res = client.try_init(&admin, &token.address, &0);
-    assert_eq!(res.unwrap_err().unwrap(), Error::AlreadyInitialized);
-
-    // State must be unchanged — campaign count must not have been reset to 0
-    assert_eq!(client.get_campaign_count(), 1);
-    assert_eq!(client.get_admin(), admin);
-    assert_eq!(client.get_token(), token.address);
-    assert_eq!(client.get_platform_fee(), 300);
-}
-
-#[test]
-fn test_update_campaign_description_rejects_empty() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Title").clone(),
-        String::from_str(&env, "Desc").clone(),
-        1_000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    let res = client.try_update_campaign_description(&campaign_id, &String::from_str(&env, ""));
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-}
-
-#[test]
-fn test_update_campaign_description_not_found() {
-    let (env, _, _, _, _, _, _, client) = setup_env();
-
-    let res = client.try_update_campaign_description(&999, &String::from_str(&env, "Some desc"));
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotFound);
-}
-
-#[test]
-fn test_campaign_ownership_transfer_flow() {
-    let (env, _admin, creator, contributor1, contributor2, _, _, client) = setup_env();
-    let new_creator = contributor1;
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Transfer Test").clone(),
-        String::from_str(&env, "Desc").clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.initiate_campaign_transfer(&campaign_id, &new_creator);
-    let campaign = client.get_campaign(&campaign_id);
-    assert_eq!(
-        campaign.pending_creator,
-        MaybePendingCreator::Some(new_creator.clone())
-    );
-    assert_eq!(campaign.creator, creator);
-
-    client.accept_campaign_transfer(&campaign_id);
-
-    let campaign_after = client.get_campaign(&campaign_id);
-    assert_eq!(campaign_after.creator, new_creator.clone());
-    assert_eq!(campaign_after.pending_creator, MaybePendingCreator::None);
-
-    let updated_description = String::from_str(&env, "Managed by the transferred owner");
-    client.update_campaign_description(&campaign_id, &updated_description);
-
-    let auths = env.auths();
-    let (auth_addr, invocation) = auths.last().unwrap();
-    assert_eq!(auth_addr, &new_creator);
-    assert_eq!(
-        invocation,
-        &AuthorizedInvocation {
-            function: AuthorizedFunction::Contract((
-                client.address.clone(),
-                Symbol::new(&env, "update_campaign_description"),
-                (campaign_id, updated_description).into_val(&env),
-            )),
-            sub_invocations: Default::default(),
-        }
-    );
-
-    let campaign_id_2 = client.create_campaign(&make_params(
-        new_creator.clone(),
-        String::from_str(&env, "Cancel Test").clone(),
-        String::from_str(&env, "Desc").clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id_2);
-    client.initiate_campaign_transfer(&campaign_id_2, &contributor2);
-    client.cancel_campaign_transfer(&campaign_id_2);
-    let final_campaign = client.get_campaign(&campaign_id_2);
-    assert_eq!(final_campaign.pending_creator, MaybePendingCreator::None);
-}
-
-#[test]
-fn test_campaign_transfer_validations() {
-    let (env, _admin, creator, contributor1, _, _, _, client) = setup_env();
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Transfer Guardrails").clone(),
-        String::from_str(&env, "Desc").clone(),
-        1000,
-        30,
-        Category::Publisher,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    let res = client.try_initiate_campaign_transfer(&campaign_id, &creator);
-    assert_eq!(res.unwrap_err().unwrap(), Error::InvalidNewOwner);
-
-    let res = client.try_accept_campaign_transfer(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NoTransferPending);
-
-    client.initiate_campaign_transfer(&campaign_id, &contributor1);
-    client.cancel_campaign_transfer(&campaign_id);
-
-    // Verify cancel was authorized by the creator
-    let auths = env.auths();
-    let (auth_addr, _) = auths.last().unwrap();
-    assert_eq!(auth_addr, &creator);
-
-    let campaign = client.get_campaign(&campaign_id);
-    assert_eq!(campaign.pending_creator, MaybePendingCreator::None);
-
-    let res = client.try_cancel_campaign_transfer(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NoTransferPending);
-}
-
-#[test]
-fn test_pause_and_unpause() {
-    let (_env, _admin, _creator, _contributor1, _, _token, _token_admin, client) = setup_env();
-
-    // Initially not paused
-    assert!(!client.is_paused());
-
-    // Pause
-    client.pause();
-
-    // Now paused
-    assert!(client.is_paused());
-
-    // Unpause
-    client.unpause();
-
-    // Not paused
-    assert!(!client.is_paused());
-}
-
-#[test]
-fn test_pause_blocks_state_changing_operations() {
-    let (env, _admin, creator, contributor1, _contributor2, token, token_admin, client) =
-        setup_env();
-
-    token_admin.mint(&contributor1, &2000);
-    token_admin.mint(&creator, &10000);
-
-    let title = String::from_str(&env, "Paused Test");
-    let desc = String::from_str(&env, "Testing pause functionality");
-
-    // Create campaign before pause
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    // Pause
-    client.pause();
-    assert!(client.is_paused());
-
-    // Try state-changing operations, should fail
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "New Campaign").clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-
-    let res = client.try_contribute(&campaign_id, &contributor1, &500);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-
-    let res = client.try_cancel_campaign(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-
-    let res = client.try_verify_campaign(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-
-    let res = client.try_update_platform_fee(&400);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-
-    // View functions should still work
-    let campaign = client.get_campaign(&campaign_id);
-    assert_eq!(campaign.title, title);
-
-    let paused = client.is_paused();
-    assert!(paused);
-
-    // Unpause
-    client.unpause();
-    assert!(!client.is_paused());
-
-    // Now operations should work
-    client.contribute(&campaign_id, &contributor1, &500);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 500);
-
-    let _ = token;
-}
-
-// ── Deadline edge-case tests (#74) ───────────────────────────────────────────
-
-/// A contribution made 1 second before the deadline must succeed.
-#[test]
-fn test_contribute_one_second_before_deadline() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Almost Deadline").clone(),
-        String::from_str(&env, "Desc").clone(),
-        1000,
-        1,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    let deadline = client.get_campaign(&campaign_id).deadline;
-
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: deadline - 1,
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence(),
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    client.contribute(&campaign_id, &contributor1, &500);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 500);
-}
-
-/// Withdrawing while the deadline has not yet passed and the goal is not met
-/// returns `FundingGoalNotReached`.
-#[test]
-fn test_withdraw_before_deadline_goal_not_met_fails() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Early Withdraw").clone(),
-        String::from_str(&env, "Desc").clone(),
-        10_000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &500);
-
-    // Deadline has not passed; goal not met — withdraw must fail
-    let res = client.try_withdraw_funds(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalNotReached);
-}
-
-/// Withdrawing after the deadline when the goal is unmet must still return the
-/// typed `FundingGoalNotReached` error.
-#[test]
-fn test_withdraw_after_deadline_goal_not_met_returns_typed_error() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Late Withdraw").clone(),
-        String::from_str(&env, "Desc").clone(),
-        10_000,
-        1,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &500);
-
-    let deadline = client.get_campaign(&campaign_id).deadline;
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: deadline + 1,
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence(),
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    let res = client.try_withdraw_funds(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalNotReached);
-}
-
-/// A contributor can claim a refund only after the deadline has passed
-/// and the campaign failed to reach its goal.
-#[test]
-fn test_refund_requires_deadline_passed_and_goal_missed() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Failed Campaign").clone(),
-        String::from_str(&env, "Desc").clone(),
-        10_000,
-        1,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &500);
-
-    // Before deadline: refund must fail (campaign still active, goal not met yet)
-    let res = client.try_claim_refund(&campaign_id, &contributor1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-
-    let deadline = client.get_campaign(&campaign_id).deadline;
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: deadline + 1,
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence(),
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    // After deadline, goal not reached: refund succeeds
-    client.claim_refund(&campaign_id, &contributor1);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 0);
-}
-
-/// A contributor whose campaign reached its goal cannot claim a refund.
-#[test]
-fn test_no_refund_when_goal_reached() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Successful Campaign").clone(),
-        String::from_str(&env, "Desc").clone(),
-        500,
-        1,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    // Meet the funding goal exactly
-    client.contribute(&campaign_id, &contributor1, &500);
-
-    let deadline = client.get_campaign(&campaign_id).deadline;
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: deadline + 1,
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence(),
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    // Goal was reached — refund must be rejected
-    let res = client.try_claim_refund(&campaign_id, &contributor1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-}
-
-#[test]
-fn test_claim_revenue_requires_contributor_auth() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &2000);
-
-    let title = String::from_str(&env, "Revenue Claim Auth");
-    let desc = String::from_str(&env, "Testing claim revenue auth");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        10,
-        Category::EducationalStartup,
-        true,
-        1000,
-        0i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &1000);
-
-    // withdraw to distribute funds
-    client.withdraw_funds(&campaign_id);
-
-    // deposit revenue
-    token_admin.mint(&creator, &5000);
-    client.deposit_revenue(&campaign_id, &5000);
-
-    // Now clear auths just in case
-    env.mock_all_auths();
-
-    client.claim_revenue(&campaign_id, &contributor1);
-
-    let auths = env.auths();
-    let found = auths.iter().any(|(addr, inv)| {
-        *addr == contributor1
-            && match &inv.function {
-                AuthorizedFunction::Contract((contract, function, _)) => {
-                    contract == &client.address && function == &Symbol::new(&env, "claim_revenue")
-                }
-                _ => false,
-            }
-    });
-
-    assert!(
-        found,
-        "contributor1 should have been authorized for claim_revenue"
-    );
-}
-
-#[test]
-fn test_set_voting_params_emits_event() {
-    let (env, admin, _creator, _contributor1, _contributor2, _token, _token_admin, client) =
-        setup_env();
-
-    client.set_voting_params(&admin, &5, &7000);
-
-    let events = env.events().all();
-    let last_event = events.last().unwrap();
-
-    // topics: (symbol, caller)
-    let topics = &last_event.1;
-    assert_eq!(topics.len(), 2);
-
-    // data: (old_quorum, new_quorum, old_threshold, new_threshold)
-    let data: (u32, u32, u32, u32) = soroban_sdk::FromVal::from_val(&env, &last_event.2);
-    assert_eq!(data, (3, 5, 6000, 7000));
-}
-
-#[test]
-fn test_list_campaigns_exclusive_cursor_semantics() {
-    let (env, _admin, creator, _c1, _c2, _token, _token_admin, client) = setup_env();
-
-    for i in 0..3 {
-        let title = String::from_str(&env, "Campaign");
-        let desc = String::from_str(&env, "Desc");
-        let id = client.create_campaign(&make_params(
-            creator.clone(),
-            title.clone(),
-            desc.clone(),
-            1000 + i as i128,
-            30,
-            Category::Learner,
-            false,
-            0,
-            0i128,
-        ));
-        assert_eq!(id, (i + 1) as u32);
-    }
-
-    let page1 = client.list_campaigns(&0, &2);
-    assert_eq!(page1.len(), 2);
-    assert_eq!(page1.get(0).unwrap().id, 1);
-    assert_eq!(page1.get(1).unwrap().id, 2);
-
-    let page2 = client.list_campaigns(&2, &2);
-    assert_eq!(page2.len(), 1);
-    assert_eq!(page2.get(0).unwrap().id, 3);
-}
-
-#[test]
-fn test_list_active_campaigns_exclusive_cursor_semantics() {
-    let (env, _admin, creator, _c1, _c2, _token, _token_admin, client) = setup_env();
-
-    for _ in 0..4 {
-        let title = String::from_str(&env, "Campaign");
-        let desc = String::from_str(&env, "Desc");
-        let _ = client.create_campaign(&make_params(
-            creator.clone(),
-            title.clone(),
-            desc.clone(),
-            1000,
-            30,
-            Category::Learner,
-            false,
-            0,
-            0i128,
-        ));
-    }
-
-    // Cancel campaign id 2 so active listing filters it out.
-    client.cancel_campaign(&2);
-
-    let active1 = client.list_active_campaigns(&0, &2);
-    assert_eq!(active1.0.len(), 2);
-    assert_eq!(active1.0.get(0).unwrap().id, 1);
-    assert_eq!(active1.0.get(1).unwrap().id, 3);
-
-    let active2 = client.list_active_campaigns(&3, &2);
-    assert_eq!(active2.0.len(), 1);
-    assert_eq!(active2.0.get(0).unwrap().id, 4);
-}
-
-#[test]
-fn test_revenue_lifecycle_e2e() {
-    let (env, _admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-
-    // Mint tokens for contributors
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&contributor2, &3000);
-
-    // Create campaign with revenue sharing enabled
-    let title = String::from_str(&env, "Revenue Sharing Campaign");
-    let desc = String::from_str(
-        &env,
-        "Full lifecycle test: create, fund, withdraw, deposit revenue, claim",
-    );
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title,
-        desc,
-        6000,
-        30,
-        Category::EducationalStartup,
-        true,
-        2000,
-        0i128,
-    ));
-
-    // Verify campaign so contributions are allowed
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    // Both contributors fund the campaign
-    client.contribute(&campaign_id, &contributor1, &4000);
-    assert_eq!(
-        client.get_contribution(&campaign_id, &contributor1),
-        4000,
-        "contributor1 contribution should be 4000"
-    );
-
-    client.contribute(&campaign_id, &contributor2, &2500);
-    assert_eq!(
-        client.get_contribution(&campaign_id, &contributor2),
-        2500,
-        "contributor2 contribution should be 2500"
-    );
-
-    // Verify campaign reached funding goal
-    let campaign = client.get_campaign(&campaign_id);
-    assert_eq!(
-        campaign.amount_raised, 6500,
-        "amount_raised should equal sum of contributions"
-    );
-    assert!(campaign.amount_raised >= campaign.funding_goal);
-
-    // Creator withdraws funds (campaign closes, distributions happen)
-    client.withdraw_funds(&campaign_id);
-
-    let campaign_after_withdrawal = client.get_campaign(&campaign_id);
-    assert!(campaign_after_withdrawal.funds_withdrawn);
-    assert!(!campaign_after_withdrawal.is_active);
-
-    // Creator deposits revenue into the pool
-    token_admin.mint(&creator, &10000);
-    client.deposit_revenue(&campaign_id, &10000);
-
-    let revenue_pool = client.get_revenue_pool(&campaign_id);
-    assert_eq!(
-        revenue_pool, 10000,
-        "revenue pool should be 10000 after deposit"
-    );
-
-    // Contributor 1 claims revenue share
-    let contrib1_claimed_before = client.get_revenue_claimed(&campaign_id, &contributor1);
-    client.claim_revenue(&campaign_id, &contributor1);
-    let contrib1_claimed_after = client.get_revenue_claimed(&campaign_id, &contributor1);
-    assert!(
-        contrib1_claimed_after > contrib1_claimed_before,
-        "contributor1 should have claimed revenue"
-    );
-
-    // Contributor 2 claims revenue share
-    let contrib2_claimed_before = client.get_revenue_claimed(&campaign_id, &contributor2);
-    client.claim_revenue(&campaign_id, &contributor2);
-    let contrib2_claimed_after = client.get_revenue_claimed(&campaign_id, &contributor2);
-    assert!(
-        contrib2_claimed_after > contrib2_claimed_before,
-        "contributor2 should have claimed revenue"
-    );
-
-    // Creator claims their retained share
-    client.claim_creator_revenue(&campaign_id);
-
-    // Verify no more revenue available for contributors
-    let result1 = client.try_claim_revenue(&campaign_id, &contributor1);
-    assert!(result1.is_err(), "second claim by contributor1 should fail");
-
-    let result2 = client.try_claim_revenue(&campaign_id, &contributor2);
-    assert!(result2.is_err(), "second claim by contributor2 should fail");
-
-    // Verify event emissions (check contribution and withdrawal events exist)
-    let events = env.events().all();
-    assert!(
-        !events.is_empty(),
-        "contract should have emitted at least one event"
-    );
-}
-
-// ── Issue #105 ────────────────────────────────────────────────────────────────
-// Boundary tests: description length values 0, 1, 1000, and 1001.
-#[test]
-fn test_description_length_boundaries() {
-    extern crate std;
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let title = String::from_str(&env, "Title");
-
-    // Length 0: must fail ValidationFailed
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        String::from_str(&env, ""),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-
-    // Length 1: must succeed (lower bound)
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        String::from_str(&env, "a"),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert!(res.is_ok(), "description of length 1 should be valid");
-
-    // Length 1000: must succeed (exactly at the upper bound)
-    let desc_1000 = "a".repeat(1000);
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        String::from_str(&env, &desc_1000),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert!(res.is_ok(), "description of length 1000 should be valid");
-
-    // Length 1001: must fail ValidationFailed (one over the upper bound)
-    let desc_1001 = "a".repeat(1001);
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        String::from_str(&env, &desc_1001),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-}
-
-#[test]
-fn test_title_length_boundaries() {
-    extern crate std;
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let desc = String::from_str(&env, "Description");
-
-    // Length 0: must fail ValidationFailed
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, ""),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-
-    // Length 1: must succeed (lower bound)
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "a"),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert!(res.is_ok(), "title of length 1 should be valid");
-
-    // Length 100: must succeed (exactly at the upper bound)
-    let title_100 = "a".repeat(100);
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, &title_100),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert!(res.is_ok(), "title of length 100 should be valid");
-
-    // Length 101: must fail ValidationFailed (one over the upper bound)
-    let title_101 = "a".repeat(101);
-    let res = client.try_create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, &title_101),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-}
-
-// ── Issue #187 ────────────────────────────────────────────────────────────────
-#[test]
-fn test_contribution_cap_persists_across_refund_recontribution_cycles() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5_000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Cap persistence"),
-        String::from_str(&env, "lifetime cap test"),
-        2_000,
-        1,
-        Category::Learner,
-        false,
-        0,
-        1_000i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-
-    client.contribute(&campaign_id, &contributor1, &900);
-
-    // Make campaign refundable, then reset current contribution via refund.
-    client.cancel_campaign(&campaign_id);
-    client.claim_refund(&campaign_id, &contributor1);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 0);
-
-    // Lifetime amount is cleared after full refund.
-    assert_eq!(
-        client.get_lifetime_contribution(&campaign_id, &contributor1),
-        0
-    );
-}
-
-// ── Issue #198 ────────────────────────────────────────────────────────────────
-#[test]
-fn test_get_campaigns_by_category_with_pagination() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let id1 = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Learner 1"),
-        String::from_str(&env, "a"),
-        100,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let _id2 = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Publisher 1"),
-        String::from_str(&env, "b"),
-        100,
-        30,
-        Category::Publisher,
-        false,
-        0,
-        0i128,
-    ));
-    let id3 = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Learner 2"),
-        String::from_str(&env, "c"),
-        100,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    let learner_page_1 = client.get_campaigns_by_category(&Category::Learner, &0, &1);
-    assert_eq!(learner_page_1.len(), 1);
-    assert_eq!(learner_page_1.get(0).unwrap().id, id1);
-
-    let learner_page_2 = client.get_campaigns_by_category(&Category::Learner, &1, &1);
-    assert_eq!(learner_page_2.len(), 1);
-    assert_eq!(learner_page_2.get(0).unwrap().id, id3);
-
-    let publisher = client.get_campaigns_by_category(&Category::Publisher, &0, &10);
-    assert_eq!(publisher.len(), 1);
-    assert_eq!(publisher.get(0).unwrap().category, Category::Publisher);
-}
-
-// ── Issue #206 ────────────────────────────────────────────────────────────────
-#[test]
-fn test_get_platform_stats_returns_aggregates() {
-    let (env, _admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-    token_admin.mint(&contributor1, &2_000);
-    token_admin.mint(&contributor2, &2_000);
-
-    let c1 = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Stats 1"),
-        String::from_str(&env, "s1"),
-        500,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    let c2 = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Stats 2"),
-        String::from_str(&env, "s2"),
-        500,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    let _ = client.try_verify_campaign(&c1);
-    let _ = client.try_verify_campaign(&c2);
-    client.contribute(&c1, &contributor1, &400);
-    client.contribute(&c2, &contributor2, &300);
-    client.cancel_campaign(&c2);
-
-    let stats = client.get_platform_stats();
-    assert_eq!(stats.total_campaigns, 2);
-    assert_eq!(stats.active_campaigns, 1);
-    assert_eq!(stats.verified_campaigns, 2);
-    assert_eq!(stats.cancelled_campaigns, 1);
-    assert_eq!(stats.total_amount_raised, 700);
-}
-
-#[test]
-fn test_total_raised_global_tracking() {
-    let (env, _admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&contributor2, &5000);
-
-    let title1 = String::from_str(&env, "Campaign 1");
-    let desc1 = String::from_str(&env, "First");
-    let c1 = client.create_campaign(&make_params(
-        creator.clone(),
-        title1.clone(),
-        desc1.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&c1);
-
-    let title2 = String::from_str(&env, "Campaign 2");
-    let desc2 = String::from_str(&env, "Second");
-    let c2 = client.create_campaign(&make_params(
-        creator.clone(),
-        title2.clone(),
-        desc2.clone(),
-        2000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&c2);
-
-    // Initial total raised should be 0
-    assert_eq!(client.get_total_raised_global(), 0);
-
-    // Contribute to C1
-    client.contribute(&c1, &contributor1, &500);
-    assert_eq!(client.get_total_raised_global(), 500);
-
-    // Contribute to C2
-    client.contribute(&c2, &contributor2, &1000);
-    assert_eq!(client.get_total_raised_global(), 1500);
-
-    // Refund C2 (Need to cancel it first or let deadline pass)
-    client.cancel_campaign(&c2);
-    client.claim_refund(&c2, &contributor2);
-    assert_eq!(client.get_total_raised_global(), 500);
-
-    // Contribute again to C1
-    client.contribute(&c1, &contributor2, &500);
-    assert_eq!(client.get_total_raised_global(), 1000);
-
-    // Withdraw C1
-    client.withdraw_funds(&c1);
-    assert_eq!(client.get_total_raised_global(), 0);
-}
-
-#[test]
-fn test_creator_campaigns_listing_and_transfer() {
-    let (env, _admin, creator1, _c1, _c2, _token, _token_admin, client) = setup_env();
-    let creator2 = Address::generate(&env);
-
-    let title1 = String::from_str(&env, "Campaign 1");
-    let desc1 = String::from_str(&env, "First");
-    let id1 = client.create_campaign(&make_params(
-        creator1.clone(),
-        title1.clone(),
-        desc1.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-
-    let title2 = String::from_str(&env, "Campaign 2");
-    let desc2 = String::from_str(&env, "Second");
-    let id2 = client.create_campaign(&make_params(
-        creator1.clone(),
-        title2.clone(),
-        desc2.clone(),
-        2000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Check creator1 list
-    let list1 = client.get_creator_campaigns(&creator1, &0, &10);
-    assert_eq!(list1.len(), 2);
-    assert_eq!(list1.get(0).unwrap().id, id1);
-    assert_eq!(list1.get(1).unwrap().id, id2);
-
-    // Test pagination
-    let paginated1 = client.get_creator_campaigns(&creator1, &0, &1);
-    assert_eq!(paginated1.len(), 1);
-    assert_eq!(paginated1.get(0).unwrap().id, id1);
-
-    let paginated2 = client.get_creator_campaigns(&creator1, &1, &1);
-    assert_eq!(paginated2.len(), 1);
-    assert_eq!(paginated2.get(0).unwrap().id, id2);
-
-    // Check creator2 list (empty)
-    let list2 = client.get_creator_campaigns(&creator2, &0, &10);
-    assert_eq!(list2.len(), 0);
-
-    // Transfer Campaign 1 to creator2
-    client.initiate_campaign_transfer(&id1, &creator2);
-    client.accept_campaign_transfer(&id1);
-
-    // Check lists after transfer
-    let list1_after = client.get_creator_campaigns(&creator1, &0, &10);
-    assert_eq!(list1_after.len(), 1);
-    assert_eq!(list1_after.get(0).unwrap().id, id2);
-
-    let list2_after = client.get_creator_campaigns(&creator2, &0, &10);
-    assert_eq!(list2_after.len(), 1);
-    assert_eq!(list2_after.get(0).unwrap().id, id1);
-}
-
-#[test]
-fn test_personal_cap_enforcement() {
-    let (env, _admin, creator, contributor1, _contributor2, _token, token_admin, client) =
-        setup_env();
-    token_admin.mint(&contributor1, &5000);
-
-    let title = String::from_str(&env, "Cap Test");
-    let desc = String::from_str(&env, "Testing caps");
-
-    // Campaign cap = 1000
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        5000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        1000i128,
-    ));
-    client.verify_campaign(&campaign_id);
-
-    // Scenario 1: Personal cap = 500 (lower than campaign cap)
-    client.set_personal_cap(&campaign_id, &contributor1, &500);
-    assert_eq!(client.get_personal_cap(&campaign_id, &contributor1), 500);
-
-    client.contribute(&campaign_id, &contributor1, &400); // OK
-    let res = client.try_contribute(&campaign_id, &contributor1, &200); // Should fail (> 500)
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContributionCapExceeded);
-
-    // Scenario 2: Personal cap = 2000 (higher than campaign cap) -> should fail validation
-    let res_set = client.try_set_personal_cap(&campaign_id, &contributor1, &2000);
-    assert_eq!(res_set.unwrap_err().unwrap(), Error::ValidationFailed);
-
-    client.set_personal_cap(&campaign_id, &contributor1, &1000);
-
-    // Total contribution currently in contract is 400.
-    // Campaign cap is 1000.
-    // Personal cap is 1000.
-    // Effective cap is min(1000, 1000) = 1000.
-    client.contribute(&campaign_id, &contributor1, &500); // Total now 900, OK
-    let res = client.try_contribute(&campaign_id, &contributor1, &200); // Should fail (> 1000)
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContributionCapExceeded);
-}
-
-#[test]
-fn test_anomaly_auto_pause_huge_contribution() {
-    let (env, _admin, creator, contributor1, _c2, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &10000);
-
-    let title = String::from_str(&env, "Science Book");
-    let desc = String::from_str(&env, "Teaching science to kids");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        2000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-
-    // Contribution > 200% of goal (2000 * 2.0 = 4000). Try 4001.
-    // In Soroban, to persist the "Paused" state, we must return Ok(()).
-    let res = client.try_contribute(&campaign_id, &contributor1, &4001);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-    // In Soroban, returning an Error rolls back state changes, including the pause.
-    assert!(!client.is_paused());
-
-    // Verify contribution was NOT recorded
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 0);
-
-    // Admin unpauses
-    client.unpause();
-    assert!(!client.is_paused());
-
-    // Normal contribution works
-    client.contribute(&campaign_id, &contributor1, &100);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 100);
-}
-
-#[test]
-fn test_anomaly_auto_pause_burst() {
-    let (env, _admin, creator, contributor1, _c2, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &10000);
-
-    let title = String::from_str(&env, "Burst Test");
-    let desc = String::from_str(&env, "Testing burst");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        2000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-
-    // Make 10 contributions in the same block (threshold is 10)
-    for _ in 0..10 {
-        client.contribute(&campaign_id, &contributor1, &10);
-    }
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 100);
-
-    // The 11th should trigger auto-pause (returns Ok for persistence)
-    let res = client.try_contribute(&campaign_id, &contributor1, &10);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-    // In Soroban, state is rolled back on Error.
-    assert!(!client.is_paused());
-
-    // Verify 11th contribution was NOT recorded
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 100);
-
-    // Admin unpauses
-    client.unpause();
-
-    // New block (ledger sequence increment) resets the counter
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: env.ledger().timestamp(),
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence() + 1,
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    client.contribute(&campaign_id, &contributor1, &10); // OK
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 110);
-}
-
-// ── Issue 4: Negative tests for deposit_revenue ──────────────────────────────
-
-#[test]
-fn test_deposit_revenue_negative_amount() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&creator, &10000);
-
-    let title = String::from_str(&env, "Startup");
-    let desc = String::from_str(&env, "Revenue sharing startup");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::EducationalStartup,
-        true,
-        2000,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.withdraw_funds(&campaign_id);
-
-    // Try to deposit negative amount
-    let res = client.try_deposit_revenue(&campaign_id, &-100);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-}
-
-#[test]
-fn test_deposit_revenue_zero_amount() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&creator, &10000);
-
-    let title = String::from_str(&env, "Startup");
-    let desc = String::from_str(&env, "Revenue sharing startup");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::EducationalStartup,
-        true,
-        2000,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.withdraw_funds(&campaign_id);
-
-    // Try to deposit zero amount
-    let res = client.try_deposit_revenue(&campaign_id, &0);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-}
-
-#[test]
-fn test_deposit_revenue_without_revenue_sharing() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&creator, &10000);
-
-    let title = String::from_str(&env, "Educator Campaign");
-    let desc = String::from_str(&env, "No revenue sharing");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.withdraw_funds(&campaign_id);
-
-    // Try to deposit revenue on non-revenue-sharing campaign
-    let res = client.try_deposit_revenue(&campaign_id, &1000);
-    assert_eq!(res.unwrap_err().unwrap(), Error::RevenueSharingNotEnabled);
-}
-
-#[test]
-fn test_deposit_revenue_when_paused() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&creator, &10000);
-
-    let title = String::from_str(&env, "Startup");
-    let desc = String::from_str(&env, "Revenue sharing startup");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::EducationalStartup,
-        true,
-        2000,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.withdraw_funds(&campaign_id);
-
-    // Pause the contract
-    client.pause();
-
-    // Try to deposit revenue when paused
-    let res = client.try_deposit_revenue(&campaign_id, &1000);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-}
-
-#[test]
-fn test_deposit_revenue_non_existent_campaign() {
-    let (_env, _admin, creator, _, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&creator, &10000);
-
-    // Try to deposit revenue for non-existent campaign
-    let res = client.try_deposit_revenue(&999, &1000);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotFound);
-}
-
-#[test]
-fn test_deposit_revenue_repeated_calls_accumulate_and_emit_events() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&creator, &10_000);
-
-    let campaign_id = client.create_campaign(&CreateCampaignParams {
-        creator: creator.clone(),
-        title: String::from_str(&env, "Repeated Deposits"),
-        description: String::from_str(&env, "Deposit idempotency"),
-        funding_goal: 1000,
-        duration_days: 30,
-        category: Category::EducationalStartup,
-        has_revenue_sharing: true,
-        revenue_share_percentage: 2000,
-        max_contribution_per_user: 0i128,
-    });
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.withdraw_funds(&campaign_id);
-
-    let events_before = env.events().all().len();
-    for _ in 0..10 {
-        client.deposit_revenue(&campaign_id, &100);
-    }
-    let events_after = env.events().all().len();
-    assert_eq!(client.get_revenue_pool(&campaign_id), 1000);
-    assert_eq!(events_after - events_before, 20);
-}
-
-// ── Issue 1: Validate refund state mutation order ────────────────────────────
-
-#[test]
-fn test_claim_refund_state_mutation_order() {
-    let (env, _admin, creator, contributor1, _, token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-
-    let title = String::from_str(&env, "Refund Order Test");
-    let desc = String::from_str(&env, "Testing state mutation order");
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        10000,
-        10,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1000);
-
-    // Cancel campaign to enable refunds
-    client.cancel_campaign(&campaign_id);
-
-    // Verify contribution exists before refund
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 1000);
-    assert_eq!(token.balance(&contributor1), 4000);
-    assert_eq!(token.balance(&client.address), 1000);
-
-    // Claim refund
-    client.claim_refund(&campaign_id, &contributor1);
-
-    // Verify state was updated correctly:
-    // 1. Contribution should be zeroed
-    assert_eq!(
-        client.get_contribution(&campaign_id, &contributor1),
-        0,
-        "contribution should be zeroed after refund"
-    );
-
-    // 2. Tokens should be transferred back
-    assert_eq!(
-        token.balance(&contributor1),
-        5000,
-        "contributor should receive refund"
-    );
-    assert_eq!(
-        token.balance(&client.address),
-        0,
-        "contract should have no balance"
-    );
-
-    // 3. Double refund should fail with NoFundsToWithdraw (not a transfer error)
-    let res = client.try_claim_refund(&campaign_id, &contributor1);
-    assert_eq!(
-        res.unwrap_err().unwrap(),
-        Error::NoFundsToWithdraw,
-        "double refund should fail with NoFundsToWithdraw, proving state was updated first"
-    );
-}
-
-#[test]
-fn test_claim_refund_multiple_contributors_isolation() {
-    let (env, _admin, creator, contributor1, contributor2, token, token_admin, client) =
-        setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&contributor2, &3000);
-
-    let title = String::from_str(&env, "Multi Refund Test");
-    let desc = String::from_str(&env, "Testing multiple refunds");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        10000,
-        10,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &2000);
-    client.contribute(&campaign_id, &contributor2, &1500);
-
-    client.cancel_campaign(&campaign_id);
-
-    // Contributor1 claims refund
-    client.claim_refund(&campaign_id, &contributor1);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 0);
-    assert_eq!(token.balance(&contributor1), 5000);
-
-    // Contributor2's state should be unaffected
-    assert_eq!(client.get_contribution(&campaign_id, &contributor2), 1500);
-    assert_eq!(token.balance(&contributor2), 1500);
-
-    // Contributor2 can still claim
-    client.claim_refund(&campaign_id, &contributor2);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor2), 0);
-    assert_eq!(token.balance(&contributor2), 3000);
-}
-
-#[test]
-fn test_claim_refund_expired_campaign() {
-    let (env, _admin, creator, contributor1, _, token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-
-    let title = String::from_str(&env, "Expired Campaign");
-    let desc = String::from_str(&env, "Will expire");
-    let duration_days = 2;
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        10000,
-        duration_days,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1000);
-
-    // Fast forward past deadline
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: env.ledger().timestamp() + (duration_days * 86450),
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence(),
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    // Should be able to claim refund for expired campaign that didn't meet goal
-    client.claim_refund(&campaign_id, &contributor1);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 0);
-    assert_eq!(token.balance(&contributor1), 5000);
-    assert_eq!(client.get_revenue_claimed(&campaign_id, &contributor1), 0);
-}
-
-// Issue #341: claim_revenue must be gated on funds_withdrawn. The prior version
-// of this test exercised a "claim before withdraw, then cancel, then refund"
-// flow which is now structurally impossible (cancel is blocked once funds are
-// withdrawn). We keep this slot as a regression test for the funds_withdrawn
-// guard.
-#[test]
-fn test_claim_revenue_blocked_before_funds_withdrawn() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&creator, &10_000);
-
-    let campaign_id = client.create_campaign(&CreateCampaignParams {
-        creator: creator.clone(),
-        title: String::from_str(&env, "Claim Gated On Withdraw"),
-        description: String::from_str(&env, "Revenue claim must wait for funds_withdrawn"),
-        funding_goal: 5000,
-        duration_days: 30,
-        category: Category::EducationalStartup,
-        has_revenue_sharing: true,
-        revenue_share_percentage: 2000,
-        max_contribution_per_user: 0i128,
-    });
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &5000);
-
-    // Funds not yet withdrawn — claim must be rejected.
-    let res = client.try_claim_revenue(&campaign_id, &contributor1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ValidationFailed);
-
-    // Revenue deposits are also blocked before withdrawal.
-    let deposit_res = client.try_deposit_revenue(&campaign_id, &1000);
-    assert_eq!(deposit_res.unwrap_err().unwrap(), Error::ValidationFailed);
-
-    // After withdrawal, deposit + claim succeeds.
-    client.withdraw_funds(&campaign_id);
-    client.deposit_revenue(&campaign_id, &1000);
-    client.claim_revenue(&campaign_id, &contributor1);
-    assert!(client.get_revenue_claimed(&campaign_id, &contributor1) > 0);
-}
-
-#[test]
-fn test_claim_refund_removes_contribution_storage_key() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5_000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Refund storage cleanup").clone(),
-        String::from_str(&env, "Contribution key should be removed").clone(),
-        5_000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1_000);
-    client.cancel_campaign(&campaign_id);
-
-    env.as_contract(&client.address, || {
-        assert!(env
-            .storage()
-            .persistent()
-            .has(&DataKey::Contribution(campaign_id, contributor1.clone())));
-    });
-
-    client.claim_refund(&campaign_id, &contributor1);
-
-    env.as_contract(&client.address, || {
-        assert!(!env
-            .storage()
-            .persistent()
-            .has(&DataKey::Contribution(campaign_id, contributor1.clone())));
-    });
-}
-
-// ── Issue 3: Fuzz/Integration tests for vote_on_campaign ─────────────────────
-
-#[test]
-fn test_vote_on_campaign_basic_flow() {
-    let (env, _admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-
-    token_admin.mint(&contributor1, &1000);
-    token_admin.mint(&contributor2, &1000);
-
-    let title = String::from_str(&env, "Voting Test");
-    let desc = String::from_str(&env, "Test voting");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Vote to approve
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(client.get_approve_votes(&campaign_id), 1);
-    assert_eq!(client.get_reject_votes(&campaign_id), 0);
-    assert!(client.has_voted(&campaign_id, &contributor1));
-
-    // Vote to reject
-    client.vote_on_campaign(&campaign_id, &contributor2, &false);
-    assert_eq!(client.get_approve_votes(&campaign_id), 1);
-    assert_eq!(client.get_reject_votes(&campaign_id), 1);
-    assert!(client.has_voted(&campaign_id, &contributor2));
-}
-
-#[test]
-fn test_vote_on_campaign_double_vote_fails() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &1000);
-
-    let title = String::from_str(&env, "Double Vote Test");
-    let desc = String::from_str(&env, "Test double voting");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-
-    // Try to vote again
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &false);
-    assert_eq!(res.unwrap_err().unwrap(), Error::AlreadyVoted);
-}
-
-#[test]
-fn test_vote_on_campaign_no_tokens_fails() {
-    let (env, _admin, creator, contributor1, _, _token, _token_admin, client) = setup_env();
-
-    let title = String::from_str(&env, "No Token Vote Test");
-    let desc = String::from_str(&env, "Test voting without tokens");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Try to vote without tokens
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NotTokenHolder);
-}
-
-#[test]
-fn test_vote_on_campaign_below_minimum_balance_fails() {
-    let (env, admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &100);
-
-    // Set minimum voting balance to 500
-    client.set_min_voting_balance(&admin, &500);
-
-    let title = String::from_str(&env, "Min Balance Vote Test");
-    let desc = String::from_str(&env, "Test voting with insufficient balance");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Try to vote with balance below minimum
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NotTokenHolder);
-}
-
-#[test]
-fn test_vote_on_verified_campaign_fails() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &1000);
-
-    let title = String::from_str(&env, "Already Verified");
-    let desc = String::from_str(&env, "Test voting on verified campaign");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Verify campaign
-    client.verify_campaign(&campaign_id);
-
-    // Try to vote on verified campaign
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignAlreadyVerified);
-}
-
-#[test]
-fn test_vote_on_cancelled_campaign_fails() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &1000);
-
-    let title = String::from_str(&env, "Cancelled Campaign");
-    let desc = String::from_str(&env, "Test voting on cancelled campaign");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Cancel campaign
-    client.cancel_campaign(&campaign_id);
-
-    // Try to vote on cancelled campaign
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotActive);
-}
-
-#[test]
-fn test_vote_on_campaign_past_deadline_fails() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &1000);
-
-    let campaign_id = client.create_campaign(&CreateCampaignParams {
-        creator: creator.clone(),
-        title: String::from_str(&env, "Deadline Vote"),
-        description: String::from_str(&env, "Voting deadline gate"),
-        funding_goal: 1000,
-        duration_days: 1,
-        category: Category::Learner,
-        has_revenue_sharing: false,
-        revenue_share_percentage: 0,
-        max_contribution_per_user: 0i128,
-    });
-
-    let deadline = client.get_campaign(&campaign_id).deadline;
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        timestamp: deadline + 1,
-        protocol_version: 22,
-        sequence_number: env.ledger().sequence(),
-        network_id: [0; 32],
-        base_reserve: 10,
-        min_temp_entry_ttl: 10,
-        min_persistent_entry_ttl: 10,
-        max_entry_ttl: 10,
-    });
-
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::DeadlinePassed);
-}
-
-#[test]
-fn test_vote_on_campaign_after_withdraw_fails() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &2000);
-
-    let campaign_id = client.create_campaign(&CreateCampaignParams {
-        creator: creator.clone(),
-        title: String::from_str(&env, "Withdrawn Vote"),
-        description: String::from_str(&env, "Voting withdrawn gate"),
-        funding_goal: 1000,
-        duration_days: 30,
-        category: Category::Learner,
-        has_revenue_sharing: false,
-        revenue_share_percentage: 0,
-        max_contribution_per_user: 0i128,
-    });
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.withdraw_funds(&campaign_id);
-
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotActive);
-}
-
-#[test]
-fn test_vote_on_campaign_token_weighted() {
-    let (env, _admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-
-    // contributor1 has more tokens
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&contributor2, &1000);
-
-    let title = String::from_str(&env, "Weighted Vote Test");
-    let desc = String::from_str(&env, "Test token-weighted voting");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-    client.vote_on_campaign(&campaign_id, &contributor2, &false);
-
-    // Both have 1 vote count
-    assert_eq!(client.get_approve_votes(&campaign_id), 1);
-    assert_eq!(client.get_reject_votes(&campaign_id), 1);
-}
-
-#[test]
-fn test_verify_campaign_with_votes_quorum_not_met() {
-    let (env, admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &1000);
-
-    // Set quorum to 5 votes
-    client.set_voting_params(&admin, &5, &6000);
-
-    let title = String::from_str(&env, "Quorum Test");
-    let desc = String::from_str(&env, "Test quorum requirement");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Only 1 vote
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-
-    // Try to verify with insufficient votes
-    let res = client.try_verify_campaign_with_votes(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::VotingQuorumNotMet);
-}
-
-#[test]
-fn test_verify_campaign_with_votes_threshold_not_met() {
-    let (env, admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-
-    token_admin.mint(&contributor1, &1000);
-    token_admin.mint(&contributor2, &1000);
-
-    let voter3 = Address::generate(&env);
-    token_admin.mint(&voter3, &1000);
-
-    // Set threshold to 80% (8000 bps)
-    client.set_voting_params(&admin, &3, &8000);
-
-    let title = String::from_str(&env, "Threshold Test");
-    let desc = String::from_str(&env, "Test approval threshold");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // 2 approve, 1 reject = 66.67% approval (below 80%)
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-    client.vote_on_campaign(&campaign_id, &contributor2, &true);
-    client.vote_on_campaign(&campaign_id, &voter3, &false);
-
-    // Try to verify with insufficient approval
-    let res = client.try_verify_campaign_with_votes(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::VotingThresholdNotMet);
-}
-
-#[test]
-fn test_verify_campaign_with_votes_success() {
-    let (env, admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-
-    token_admin.mint(&contributor1, &1000);
-    token_admin.mint(&contributor2, &1000);
-
-    let voter3 = Address::generate(&env);
-    token_admin.mint(&voter3, &1000);
-
-    // Set quorum to 3, threshold to 60%
-    client.set_voting_params(&admin, &3, &6000);
-
-    let title = String::from_str(&env, "Success Verify Test");
-    let desc = String::from_str(&env, "Test successful verification");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // 2 approve, 1 reject = 66.67% approval (above 60%)
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-    client.vote_on_campaign(&campaign_id, &contributor2, &true);
-    client.vote_on_campaign(&campaign_id, &voter3, &false);
-
-    // Should verify successfully
-    client.verify_campaign_with_votes(&campaign_id);
-
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(campaign.is_verified);
-}
-
-#[test]
-fn test_vote_on_nonexistent_campaign() {
-    let (_env, _admin, _creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &1000);
-
-    // Try to vote on non-existent campaign
-    let res = client.try_vote_on_campaign(&999, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotFound);
-}
-
-#[test]
-fn test_min_voting_balance_threshold_enforcement() {
-    let (env, admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-
-    // Mint different token amounts
-    token_admin.mint(&contributor1, &50); // Below threshold
-    token_admin.mint(&contributor2, &200); // Above threshold
-
-    let title = String::from_str(&env, "Min Balance Vote Test");
-    let desc = String::from_str(&env, "Testing minimum voting balance");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title.clone(),
-        desc.clone(),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Set minimum voting balance to 100 tokens
-    client.set_min_voting_balance(&admin, &100);
-    assert_eq!(client.get_min_voting_balance(), 100);
-
-    // contributor1 (50 tokens) should be rejected
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NotTokenHolder);
-
-    // contributor2 (200 tokens) should succeed
-    client.vote_on_campaign(&campaign_id, &contributor2, &true);
-    assert!(client.has_voted(&campaign_id, &contributor2));
-    assert_eq!(client.get_approve_votes(&campaign_id), 1);
-
-    // Admin can update threshold to 0 (no restriction)
-    client.set_min_voting_balance(&admin, &0);
-    assert_eq!(client.get_min_voting_balance(), 0);
-
-    // Now contributor1 can vote
-    client.vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert!(client.has_voted(&campaign_id, &contributor1));
-    assert_eq!(client.get_approve_votes(&campaign_id), 2);
-}
-
-// ── Tests for #169: withdraw_funds requires is_verified ─────────────────────
-
-#[test]
-fn test_withdraw_funds_requires_verified_campaign() {
-    let (env, _admin, creator, _contributor1, _, _token, token_admin, client) = setup_env();
-
-    let title = String::from_str(&env, "Unverified Campaign");
-    let desc = String::from_str(&env, "Description");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title,
-        desc,
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Seed the contract with tokens and manipulate storage directly to simulate
-    // a funded but unverified campaign — the defense-in-depth guard in
-    // withdraw_funds must catch this even though contribute also enforces it.
-    let contract_id = env.register_contract(None, ProofOfHeart);
-    token_admin.mint(&contract_id, &1500);
-    env.as_contract(&client.address, || {
-        let mut campaign = storage::get_campaign(&env, campaign_id).unwrap();
-        campaign.amount_raised = 1500;
-        // Explicitly keep is_verified = false
-        storage::set_campaign(&env, campaign_id, &campaign);
-    });
-
-    // withdraw_funds should fail because campaign is not verified
-    let result = client.try_withdraw_funds(&campaign_id);
-    assert_eq!(result.unwrap_err().unwrap(), Error::CampaignNotVerified);
-}
-
-#[test]
-fn test_withdraw_funds_succeeds_when_verified() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5000);
-
-    let title = String::from_str(&env, "Verified Campaign");
-    let desc = String::from_str(&env, "Description");
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title,
-        desc,
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1500);
-
-    // Should succeed now that campaign is verified
-    assert!(client.try_withdraw_funds(&campaign_id).is_ok());
-}
-
-// ── Tests for #188: revenue_share_percentage normalised to 0 ────────────────
-
-#[test]
-fn test_revenue_share_percentage_normalised_to_zero_when_disabled() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let title = String::from_str(&env, "No Revenue Campaign");
-    let desc = String::from_str(&env, "Description");
-    // Pass a non-zero percentage with has_revenue_sharing=false
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title,
-        desc,
-        1000,
-        30,
-        Category::Educator,
-        false,
-        12345,
-        0i128,
-    ));
-
-    let campaign = client.get_campaign(&campaign_id);
-    // The stored percentage must be 0 regardless of what was passed
-    assert_eq!(campaign.revenue_share_percentage, 0);
-    assert!(!campaign.has_revenue_sharing);
-}
-
-#[test]
-fn test_revenue_share_above_max_rejected_even_without_flag() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let title = String::from_str(&env, "Bad Revenue");
-    let desc = String::from_str(&env, "Description");
-    // Pass a percentage above REVENUE_SHARE_MAX_BPS (5000) with flag=false
-    // After normalisation to 0, the > REVENUE_SHARE_MAX_BPS check passes (0 <= 5000).
-    // This confirms the normalisation happens before the validation.
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        title,
-        desc,
-        1000,
-        30,
-        Category::Educator,
-        false,
-        9999,
-        0i128,
-    ));
-    let campaign = client.get_campaign(&campaign_id);
-    assert_eq!(campaign.revenue_share_percentage, 0);
-}
-
-#[test]
-fn test_revenue_share_with_flag_true_above_max_rejected() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let title = String::from_str(&env, "Too High Revenue");
-    let desc = String::from_str(&env, "Description");
-    let result = client.try_create_campaign(&make_params(
-        creator.clone(),
-        title,
-        desc,
-        1000,
-        30,
-        Category::EducationalStartup,
-        true,
-        5001,
-        0i128,
-    ));
-    assert_eq!(result.unwrap_err().unwrap(), Error::InvalidRevenueShare);
-}
-
-// ── Issue #185: cancel_campaign terminal-state coverage ──────────────────────
-
-/// Cancelling an already-cancelled campaign must fail with CampaignNotActive.
-#[test]
-fn test_cancel_campaign_already_cancelled_is_terminal() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Terminal Test"),
-        String::from_str(&env, "Already cancelled"),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    client.cancel_campaign(&campaign_id);
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(campaign.is_cancelled);
-    assert!(!campaign.is_active);
-
-    // Cancelling again must be rejected.
-    let res = client.try_cancel_campaign(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotActive);
-}
-
-/// Cancelling a campaign whose funds have already been withdrawn must fail
-/// with CampaignNotActive (is_active is false after withdrawal).
-#[test]
-fn test_cancel_campaign_after_withdrawal_is_terminal() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &2000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Withdrawal Terminal"),
-        String::from_str(&env, "Funds already out"),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1000);
-    client.withdraw_funds(&campaign_id);
-
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(campaign.funds_withdrawn);
-    assert!(!campaign.is_active);
-
-    // Attempting to cancel a withdrawn campaign must be rejected.
-    let res = client.try_cancel_campaign(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotActive);
-}
-
-// ── Issue Fix Verifications ──────────────────────────────────────────────────
-
-#[test]
-fn test_update_description_after_contribution() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &1000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Title"),
-        String::from_str(&env, "Old Description"),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &500);
-
-    let new_desc = String::from_str(&env, "New Description After Contribution");
-    client.update_campaign_description(&campaign_id, &new_desc);
-
-    let campaign = client.get_campaign(&campaign_id);
-    assert_eq!(campaign.description, new_desc);
-}
-
-#[test]
-fn test_update_campaign_with_contributions_fails() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &1000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Title"),
-        String::from_str(&env, "Old Description"),
-        1000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &500);
-
-    let new_title = String::from_str(&env, "New Title");
-    let new_desc = String::from_str(&env, "New Description");
-    let res = client.try_update_campaign(&campaign_id, &new_title, &new_desc);
-
-    // update_campaign is blocked after verification (CampaignAlreadyVerified takes
-    // precedence over the amount_raised > 0 check since it's checked first).
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignAlreadyVerified);
-}
-
-#[test]
-fn test_create_campaign_validation_independence() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    // Set a category cap of 10 days
-    env.as_contract(&client.address, || {
-        set_category_duration_cap(&env, Category::Educator, 10);
-    });
-
-    // 1. FundingGoalTooHigh should trigger even if duration is invalid
-    // Provide duration = 11 (invalid for Educator) and goal > max
-    let params = make_params(
-        creator.clone(),
-        String::from_str(&env, "Title"),
-        String::from_str(&env, "Desc"),
-        CAMPAIGN_FUNDING_GOAL_MAX + 1,
-        11,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    );
-
-    // Current logic checks goal bounds FIRST, then duration.
-    // Wait, let's check src/lib.rs order.
-    // 222: if funding_goal <= 0 ...
-    // 225: if funding_goal < min ...
-    // 228: let duration_max = ...
-    // 230: if !(min..=max).contains(&duration_days) { return Err(InvalidDuration); }
-    // 233: if funding_goal > get_max_campaign_funding_goal(...) { return Err(FundingGoalTooHigh); }
-
-    // In my current version, InvalidDuration (230) is checked BEFORE FundingGoalTooHigh (233).
-    // The user's requested fix for Issue 4 says:
-    /*
-    if !(CAMPAIGN_DURATION_MIN_DAYS..=duration_max).contains(&duration_days) {
-        return Err(Error::InvalidDuration);
-    }
-    if funding_goal > get_max_campaign_funding_goal(&env, CAMPAIGN_FUNDING_GOAL_MAX) {
-        return Err(Error::FundingGoalTooHigh);
-    }
-    */
-    // This is exactly what I have in src/lib.rs.
-    // But the user's Acceptance says:
-    // "FundingGoalTooHigh triggers regardless of duration validity"
-
-    // Wait! If they want FundingGoalTooHigh to trigger REGARDLESS of duration validity,
-    // it MUST be checked BEFORE duration validity.
-
-    let res = client.try_create_campaign(&params);
-    // FundingGoalTooHigh triggers regardless of duration validity (as requested).
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalTooHigh);
-
-    // 2. High goal with valid duration should trigger FundingGoalTooHigh
-    let params_valid_dur = make_params(
-        creator.clone(),
-        String::from_str(&env, "Title"),
-        String::from_str(&env, "Desc"),
-        CAMPAIGN_FUNDING_GOAL_MAX + 1,
-        5,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    );
-    let res = client.try_create_campaign(&params_valid_dur);
-    assert_eq!(res.unwrap_err().unwrap(), Error::FundingGoalTooHigh);
-}
-
-// ── Issue #260: deposit_revenue rejects cancelled campaigns ──────────────────
-
-#[test]
-fn test_deposit_revenue_cancelled_campaign() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&creator, &10000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Startup"),
-        String::from_str(&env, "Revenue sharing startup"),
-        1000,
-        30,
-        Category::EducationalStartup,
-        true,
-        2000,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-
-    // Cancel the campaign (before withdrawal — cancellation after withdrawal is disallowed)
-    client.cancel_campaign(&campaign_id);
-
-    // Depositing revenue into a cancelled campaign should fail
-    let res = client.try_deposit_revenue(&campaign_id, &500);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotActive);
-}
-
-// ── Issue #261: claim_refund clears LifetimeContribution ────────────────────
-
-#[test]
-fn test_claim_refund_clears_lifetime_contribution() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "LT Cleanup"),
-        String::from_str(&env, "Test lifetime cleanup"),
-        2000,
-        1,
-        Category::Learner,
-        false,
-        0,
-        1_000i128,
-    ));
-    let _ = client.try_verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &900);
-
-    // Cancel and refund
-    client.cancel_campaign(&campaign_id);
-    client.claim_refund(&campaign_id, &contributor1);
-
-    // LifetimeContribution should be cleared after full refund
-    assert_eq!(
-        client.get_lifetime_contribution(&campaign_id, &contributor1),
-        0,
-        "LifetimeContribution should be 0 after full refund"
-    );
-}
-
-// ── Issue #262: get_platform_stats reflects only held funds ──────────────────
-
-#[test]
-fn test_platform_stats_after_withdrawal() {
-    let (env, _admin, creator, contributor1, contributor2, _token, token_admin, client) =
-        setup_env();
-    token_admin.mint(&contributor1, &5000);
-    token_admin.mint(&contributor2, &5000);
-
-    // Campaign 1: fund and withdraw
-    let c1 = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Withdrawn"),
-        String::from_str(&env, "w"),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&c1);
-    client.contribute(&c1, &contributor1, &1000);
-    client.withdraw_funds(&c1);
-
-    // Campaign 2: still active, funded
-    let c2 = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Active"),
-        String::from_str(&env, "a"),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&c2);
-    client.contribute(&c2, &contributor2, &500);
-
-    let stats = client.get_platform_stats();
-    // Only currently held funds (campaign 2's 500), not the withdrawn 1000
-    assert_eq!(stats.total_amount_raised, 500);
-}
-
-// ── Issue #263: verify_campaigns extends voting state TTL ────────────────────
-
-#[test]
-fn test_verify_campaigns_extends_voting_state_ttl() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    // Create a campaign
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "TTL Test"),
-        String::from_str(&env, "Testing TTL extension"),
-        1000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // Bulk verify the campaign
-    let count = client.verify_campaigns(&soroban_sdk::Vec::from_array(&env, [campaign_id]));
-    assert_eq!(count, 1);
-
-    // Verify campaign is verified (confirming it worked)
-    let campaign = client.get_campaign(&campaign_id);
-    assert!(campaign.is_verified);
-}
-
-// Issue #300: verify no panics when claim_revenue / claim_creator_revenue are called
-// after all contributors have been refunded (contribution entries removed from storage).
-#[test]
-fn test_revenue_claim_after_full_refunds_no_panic() {
-    let (env, _admin, creator, contributor1, _, token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5_000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Revenue Refund Test"),
-        String::from_str(&env, "Testing revenue claim after full refund"),
-        1_000,
-        30,
-        Category::EducationalStartup,
-        true,
-        2000, // 20% revenue share
-        0i128,
-    ));
-
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &500);
-
-    // Cancel so the contributor is eligible for a refund
-    client.cancel_campaign(&campaign_id);
-
-    // Contributor claims full refund — removes their contribution entry
-    client.claim_refund(&campaign_id, &contributor1);
-    assert_eq!(token.balance(&contributor1), 5_000);
-
-    // claim_revenue must not panic; campaign is cancelled so expect CampaignNotActive
-    let res = client.try_claim_revenue(&campaign_id, &contributor1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::CampaignNotActive);
-
-    // claim_creator_revenue must not panic; no revenue deposited so expect NoFundsToWithdraw
-    let res = client.try_claim_creator_revenue(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::NoFundsToWithdraw);
-}
-
-// Directly exercises the AmountRaisedIsZero guard in claim_revenue by forcing
-// amount_raised = 0 on the campaign struct while a contribution is still present.
-// This is an artificial state that cannot arise in normal contract flow, but it
-// confirms the guard fires and returns an error rather than dividing by zero.
-#[test]
-fn test_claim_revenue_amount_raised_zero_guard() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5_000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Zero Raised Guard"),
-        String::from_str(&env, "Directly test AmountRaisedIsZero guard"),
-        1_000,
-        30,
-        Category::EducationalStartup,
-        true,
-        2000,
-        0i128,
-    ));
-
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &500);
-
-    // Artificially zero out amount_raised and effective_amount_raised while keeping
-    // the contribution in storage. Also force funds_withdrawn=true so we bypass the
-    // funds_withdrawn guard and exercise the AmountRaisedIsZero guard specifically.
-    env.as_contract(&client.address, || {
-        let mut campaign = get_campaign(&env, campaign_id).unwrap();
-        campaign.amount_raised = 0;
-        campaign.effective_amount_raised = 0;
-        campaign.funds_withdrawn = true;
-        set_campaign(&env, campaign_id, &campaign);
-    });
-
-    let res = client.try_claim_revenue(&campaign_id, &contributor1);
-    assert_eq!(res.unwrap_err().unwrap(), Error::AmountRaisedIsZero);
-}
-
-// ── Issue #379: cast_vote returns DeadlinePassed after deadline ───────────────
-
-#[test]
-fn test_vote_on_campaign_after_deadline_returns_deadline_passed() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &500);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Deadline Vote Test"),
-        String::from_str(&env, "Voting after deadline must return DeadlinePassed"),
-        1_000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    let campaign = client.get_campaign(&campaign_id);
-
-    // Advance past the deadline
-    env.ledger().with_mut(|li| {
-        li.timestamp = campaign.deadline + 1;
-    });
-
-    let res = client.try_vote_on_campaign(&campaign_id, &contributor1, &true);
-    assert_eq!(res.unwrap_err().unwrap(), Error::DeadlinePassed);
-}
-
-// ── Issue #378: verify_campaigns returns Err on partial failure ───────────────
-
-#[test]
-fn test_verify_campaigns_partial_failure_returns_err() {
-    let (env, _admin, creator, _, _, _, _, client) = setup_env();
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Valid Campaign"),
-        String::from_str(&env, "One valid campaign"),
-        1_000,
-        30,
-        Category::Learner,
-        false,
-        0,
-        0i128,
-    ));
-
-    // 999 does not exist — will produce CampaignNotFound
-    let ids = soroban_sdk::Vec::from_array(&env, [campaign_id, 999u32]);
-    let res = client.try_verify_campaigns(&ids);
-    assert!(res.unwrap_err().is_ok()); // Err variant, inner Ok means contract error
-}
-
-#[test]
-fn test_huge_contribution_triggers_auto_pause() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-
-    let campaign_id = client.create_campaign(&CreateCampaignParams {
-        creator: creator.clone(),
-        title: String::from_str(&env, "Huge Contribution Test"),
-        description: String::from_str(&env, "Testing auto-pause via huge contribution"),
-        funding_goal: 1000,
-        duration_days: 30,
-        category: Category::Learner,
-        has_revenue_sharing: false,
-        revenue_share_percentage: 0,
-        max_contribution_per_user: 0,
-    });
-    client.verify_campaign(&campaign_id);
-
-    // Anomaly detection fires (the Err rollback means AutoPaused doesn't persist
-    // through contribute() itself — test the detection, not the persistence).
-    let res = client.try_contribute(&campaign_id, &contributor1, &2001i128);
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-}
-
-#[test]
-fn test_unpause_clears_auto_pause_when_resume_campaign_blocked() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-
-    token_admin.mint(&contributor1, &5000);
-
-    let campaign_id = client.create_campaign(&CreateCampaignParams {
-        creator: creator.clone(),
-        title: String::from_str(&env, "Unpause Recovery Test"),
-        description: String::from_str(&env, "Testing unpause when resume_campaign is blocked"),
-        funding_goal: 1000,
-        duration_days: 30,
-        category: Category::Learner,
-        has_revenue_sharing: false,
-        revenue_share_percentage: 0,
-        max_contribution_per_user: 0,
-    });
-    client.verify_campaign(&campaign_id);
-
-    // Set AutoPaused directly (Soroban rolls back writes on Err, so we can't
-    // rely on the anomaly trigger in contribute() to persist the flag).
-    env.as_contract(&client.address, || {
-        env.storage().instance().set(&DataKey::AutoPaused, &true);
-    });
-
-    // Operations are blocked while AutoPaused is set
-    let res = client.try_create_campaign(&CreateCampaignParams {
-        creator: creator.clone(),
-        title: String::from_str(&env, "Should Fail"),
-        description: String::from_str(&env, "Desc"),
-        funding_goal: 500,
-        duration_days: 30,
-        category: Category::Learner,
-        has_revenue_sharing: false,
-        revenue_share_percentage: 0,
-        max_contribution_per_user: 0,
-    });
-    assert_eq!(res.unwrap_err().unwrap(), Error::ContractPaused);
-
-    // unpause() clears both Paused and AutoPaused
-    client.unpause();
-
-    // Now operations work again
-    client.contribute(&campaign_id, &contributor1, &500i128);
-    assert_eq!(client.get_contribution(&campaign_id, &contributor1), 500);
-
-    // Cancel the campaign (was blocked while auto-paused)
-    client.cancel_campaign(&campaign_id);
-
-    // resume_campaign fails because campaign is cancelled
-    let res2 = client.try_resume_campaign(&campaign_id, &creator);
-    assert_eq!(res2.unwrap_err().unwrap(), Error::CampaignNotActive);
-
-    // But operations still work because unpause already cleared AutoPaused
-    let new_id = client.create_campaign(&CreateCampaignParams {
-        creator: creator.clone(),
-        title: String::from_str(&env, "Recovered"),
-        description: String::from_str(&env, "Should work now"),
-        funding_goal: 500,
-        duration_days: 30,
-        category: Category::Learner,
-        has_revenue_sharing: false,
-        revenue_share_percentage: 0,
-        max_contribution_per_user: 0,
-    });
-    assert!(new_id > 1);
-}
-
-// ── Issue #408: checked arithmetic on the remaining fee/revenue multiplications.
-//    A pathological amount must yield Error::Overflow, never a panic. ────────────
-
-#[test]
-fn test_withdraw_funds_overflow_returns_error_not_panic() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5_000);
-
-    // 10% fee so `amount_raised * fee_bps` can overflow at extreme values.
-    client.update_platform_fee(&1000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Withdraw Overflow"),
-        String::from_str(&env, "amount_raised * fee must not panic"),
-        1_000,
-        30,
-        Category::Educator,
-        false,
-        0,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1_000);
-
-    // Force a pathological amount_raised that overflows the fee multiplication.
-    env.as_contract(&client.address, || {
-        let mut campaign = get_campaign(&env, campaign_id).unwrap();
-        campaign.amount_raised = i128::MAX;
-        set_campaign(&env, campaign_id, &campaign);
-    });
-
-    let res = client.try_withdraw_funds(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::Overflow);
-}
-
-#[test]
-fn test_claim_creator_revenue_overflow_returns_error_not_panic() {
-    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
-    token_admin.mint(&contributor1, &5_000);
-
-    let campaign_id = client.create_campaign(&make_params(
-        creator.clone(),
-        String::from_str(&env, "Creator Revenue Overflow"),
-        String::from_str(&env, "total_pool * share must not panic"),
-        1_000,
-        30,
-        Category::EducationalStartup,
-        true,
-        5000,
-        0i128,
-    ));
-    client.verify_campaign(&campaign_id);
-    client.contribute(&campaign_id, &contributor1, &1_000);
-
-    // Inject a pathological revenue pool that overflows total_pool * share_bps.
-    env.as_contract(&client.address, || {
-        set_revenue_pool(&env, campaign_id, i128::MAX);
-    });
-
-    let res = client.try_claim_creator_revenue(&campaign_id);
-    assert_eq!(res.unwrap_err().unwrap(), Error::Overflow);
-}
