+++ src/tests/mod.rs
@@ -1,23 +1,30 @@
 pub(crate) mod helpers;
 
 mod test_admin;
+mod test_benchmark;
 mod test_campaign_create;
 mod test_campaign_update;
 mod test_cancel_revenue_orphan;
 mod test_contribute;
 mod test_contribute_caps;
+mod test_create_campaign_proptest;
 mod test_creator_buckets;
 mod test_deadline_ext;
 mod test_duration_cap;
 mod test_fee_override;
 mod test_init;
+mod test_issues;
+mod test_lifecycle_events;
 mod test_listing;
 mod test_purge_voting;
 mod test_refund;
 mod test_refund_edge;
 mod test_revenue;
 mod test_revenue_deposit;
+mod test_revenue_share_proptest;
 mod test_storage_ttl;
+mod test_vesting;
 mod test_voting;
+mod test_voting_proptest;
 mod test_voting_verify;
 mod test_withdraw;
