+++ src/tests/test_issues.rs
@@ -1,19 +1,15 @@
-use super::*;
-use soroban_sdk::{
-    testutils::Address as _, testutils::Events as _, testutils::Ledger, Address, Env,
-};
-
-use crate::test::setup_env;
+use super::helpers::*;
+use crate::{storage, Campaign, DataKey, Error, MaybePendingCreator};
+use soroban_sdk::{Address, Env, String};
 
 // ── #266 migrate ──────────────────────────────────────────────────────────────
 
 #[test]
 fn test_migrate_success() {
     let (_env, admin, _, _, _, _, _, client) = setup_env();
-    // version is 1 after init; migrate from 1 → CONTRACT_VERSION (1)
     let result = client.try_migrate(&admin, &1u32);
     assert!(result.is_ok());
-    assert_eq!(client.get_version(), 1u32); // CONTRACT_VERSION = 1
+    assert_eq!(client.get_version(), 1u32);
 }
 
 #[test]
@@ -30,7 +26,6 @@ fn test_migrate_double_run_fails() {
         env.storage().instance().set(&DataKey::Version, &0u32);
     });
     client.migrate(&admin, &0u32);
-    // version is now CONTRACT_VERSION; calling again with old version fails
     let result = client.try_migrate(&admin, &0u32);
     assert_eq!(result.unwrap_err().unwrap(), Error::ValidationFailed);
 }
@@ -54,7 +49,6 @@ fn test_propose_token_update_stores_pending() {
     let (env, admin, _, _, _, _, _, client) = setup_env();
     let new_token = setup_second_token(&env, &admin);
     client.propose_token_update(&admin, &new_token);
-    // pending token is set; original token unchanged
     assert_ne!(client.get_token(), new_token);
 }
 
@@ -63,7 +57,6 @@ fn test_accept_token_update_before_delay_fails() {
     let (env, admin, _, _, _, _, _, client) = setup_env();
     let new_token = setup_second_token(&env, &admin);
     client.propose_token_update(&admin, &new_token);
-    // try to accept immediately — should fail
     let result = client.try_accept_token_update(&admin);
     assert_eq!(result.unwrap_err().unwrap(), Error::ValidationFailed);
 }
@@ -74,7 +67,6 @@ fn test_accept_token_update_after_delay_succeeds() {
     let new_token = setup_second_token(&env, &admin);
     client.propose_token_update(&admin, &new_token);
 
-    // advance time by 7 days + 1 second
     env.ledger().with_mut(|l| {
         l.timestamp += 7 * 86400 + 1;
     });
@@ -90,7 +82,6 @@ fn test_cancel_token_update_clears_pending() {
     client.propose_token_update(&admin, &new_token);
     client.cancel_token_update(&admin);
 
-    // after cancel, accept should fail (no pending)
     let result = client.try_accept_token_update(&admin);
     assert_eq!(result.unwrap_err().unwrap(), Error::ValidationFailed);
 }
@@ -116,8 +107,8 @@ fn test_propose_token_update_non_admin_fails() {
 fn make_campaign_params_simple(env: &Env, creator: &Address) -> CreateCampaignParams {
     CreateCampaignParams {
         creator: creator.clone(),
-        title: soroban_sdk::String::from_str(env, "T"),
-        description: soroban_sdk::String::from_str(env, "D"),
+        title: String::from_str(env, "T"),
+        description: String::from_str(env, "D"),
         funding_goal: 1,
         duration_days: 30,
         category: Category::Learner,
@@ -155,7 +146,6 @@ fn test_platform_stats_after_verify() {
     client.verify_campaign(&id);
     let stats = client.get_platform_stats();
     assert_eq!(stats.verified_campaigns, 1);
-    // campaign is still active after verify
     assert_eq!(stats.active_campaigns, 1);
 }
 
@@ -166,10 +156,8 @@ fn test_platform_stats_after_withdraw() {
     client.verify_campaign(&id);
 
     token_admin.mint(&contributor, &1000);
-    // funding_goal is 1, contribute 1 to meet it
     client.contribute(&id, &contributor, &1);
 
-    // advance past deadline
     env.ledger().with_mut(|l| {
         l.timestamp += 31 * 86400;
     });
@@ -185,12 +173,10 @@ fn test_platform_stats_after_withdraw() {
 fn test_get_campaigns_by_category_capped_at_list_max_limit() {
     let (env, _, creator, _, _, _, _, client) = setup_env();
 
-    // Create 60 campaigns in the Learner category
     for _ in 0..60 {
         client.create_campaign(&make_campaign_params_simple(&env, &creator));
     }
 
-    // Request 1000 — should be capped at LIST_MAX_LIMIT (50)
     let result = client.get_campaigns_by_category(&Category::Learner, &0u32, &1000u32);
     assert!(result.len() <= 50);
     assert_eq!(result.len(), 50);
@@ -243,11 +229,9 @@ fn test_paused_admin_parameter_setting_functions_fail() {
 
     client.pause();
 
-    // set_campaign_fee_override should fail when paused
     let result_fee = client.try_set_campaign_fee_override(&admin, &campaign_id, &100u32);
     assert_eq!(result_fee.unwrap_err().unwrap(), Error::ContractPaused);
 
-    // set_creation_disabled should fail when paused
     let result_disabled = client.try_set_creation_disabled(&true);
     assert_eq!(result_disabled.unwrap_err().unwrap(), Error::ContractPaused);
 }
@@ -257,11 +241,10 @@ fn test_paused_admin_parameter_setting_functions_fail() {
 fn test_set_personal_cap_cannot_exceed_max_contribution_per_user() {
     let (env, _, creator, contributor, _, _, _, client) = setup_env();
 
-    // Create a campaign with a max_contribution_per_user cap of 500
     let params = CreateCampaignParams {
         creator: creator.clone(),
-        title: soroban_sdk::String::from_str(&env, "T"),
-        description: soroban_sdk::String::from_str(&env, "D"),
+        title: String::from_str(&env, "T"),
+        description: String::from_str(&env, "D"),
         funding_goal: 1000,
         duration_days: 30,
         category: Category::Learner,
@@ -271,11 +254,9 @@ fn test_set_personal_cap_cannot_exceed_max_contribution_per_user() {
     };
     let campaign_id = client.create_campaign(&params);
 
-    // Setting personal cap equal to or less than 500 should succeed
     let res1 = client.try_set_personal_cap(&campaign_id, &contributor, &500);
     assert!(res1.is_ok());
 
-    // Setting personal cap greater than 500 should fail
     let res2 = client.try_set_personal_cap(&campaign_id, &contributor, &501);
     assert_eq!(res2.unwrap_err().unwrap(), Error::ValidationFailed);
 }
@@ -286,27 +267,22 @@ fn test_vote_weight_overflow_fails() {
     let (env, _admin, creator, contributor, _, _token, token_admin, client) = setup_env();
     let campaign_id = client.create_campaign(&make_campaign_params_simple(&env, &creator));
 
-    // Mint contributor tokens
     token_admin.mint(&contributor, &1000);
 
-    // Set high weight manually in storage to simulate a whale or accumulation that would overflow i128::MAX
     env.as_contract(&client.address, || {
         env.storage()
             .persistent()
             .set(&DataKey::ApproveWeight(campaign_id), &(i128::MAX - 500));
     });
 
-    // Cast a vote with balance 501, which overflows i128::MAX when added to i128::MAX - 500
     token_admin.mint(&contributor, &501);
 
-    // cast vote should return Overflow error
     let res = client.try_vote_on_campaign(&campaign_id, &contributor, &true);
     assert_eq!(res.unwrap_err().unwrap(), Error::Overflow);
 }
 
 // ── #360 resume_campaign admin-path coverage ──────────────────────────────────
 
-/// Helper: set the contract into auto-paused state directly in storage.
 fn set_auto_paused(env: &Env, client_address: &Address, paused: bool) {
     env.as_contract(client_address, || {
         env.storage().instance().set(&DataKey::AutoPaused, &paused);
@@ -321,7 +297,6 @@ fn test_resume_by_admin() {
     set_auto_paused(&env, &client.address, true);
     assert!(client.is_paused());
 
-    // Admin (not the creator) should be able to resume.
     client.resume_campaign(&campaign_id, &admin);
     assert!(!client.is_paused());
 }
@@ -350,11 +325,9 @@ fn test_resume_after_campaign_transfer_uses_new_creator() {
     set_auto_paused(&env, &client.address, true);
     assert!(client.is_paused());
 
-    // New creator can resume; original creator can no longer.
     client.resume_campaign(&campaign_id, &new_creator);
     assert!(!client.is_paused());
 
-    // Re-pause and verify the original creator is now rejected.
     set_auto_paused(&env, &client.address, true);
     let result = client.try_resume_campaign(&campaign_id, &original_creator);
     assert_eq!(result.unwrap_err().unwrap(), Error::NotAuthorized);
