+++ src/tests/test_refund_edge.rs
@@ -1,6 +1,6 @@
 use super::helpers::*;
 use crate::storage;
-use crate::{Category, Error};
+use crate::{Category, DataKey, Error};
 use soroban_sdk::String;
 
 #[test]
@@ -213,7 +213,158 @@ fn test_claim_revenue_after_single_refund_uses_live_raised() {
     assert_eq!(token.balance(&contributor2), 4500);
     assert_eq!(client.get_revenue_claimed(&campaign_id, &contributor2), 500);
 }
+fn has_persistent_key(env: &Env, client: &ProofOfHeartClient<'_>, key: DataKey) -> bool {
+    env.as_contract(&client.address, || env.storage().persistent().has(&key))
+}
+
+#[test]
+fn test_storage_cleaned_after_claim_refund_on_cancel() {
+    let (env, _admin, creator, contributor1, _contributor2, _token, token_admin, client) =
+        setup_env();
+
+    token_admin.mint(&contributor1, &5000);
+
+    let id = client.create_campaign(&make_params(
+        creator.clone(),
+        String::from_str(&env, "Cleanup Cancel Refund"),
+        String::from_str(&env, "Storage cleanup test"),
+        10_000,
+        30,
+        Category::Learner,
+        false,
+        0,
+        0i128,
+    ));
+    client.verify_campaign(&id);
+    client.contribute(&id, &contributor1, &1000);
+
+    client.cancel_campaign(&id);
+    client.claim_refund(&id, &contributor1);
+
+    assert!(
+        !has_persistent_key(
+            &env,
+            &client,
+            DataKey::Contribution(id, contributor1.clone())
+        ),
+        "Contribution key must be removed after refund"
+    );
+    assert!(
+        !has_persistent_key(
+            &env,
+            &client,
+            DataKey::RevenueClaimed(id, contributor1.clone())
+        ),
+        "RevenueClaimed key must not exist after refund"
+    );
+}
+
+#[test]
+fn test_storage_cleaned_after_claim_refund_on_failed_campaign() {
+    let (env, _admin, creator, contributor1, _contributor2, _token, token_admin, client) =
+        setup_env();
+
+    token_admin.mint(&contributor1, &5000);
+
+    let id = client.create_campaign(&make_params(
+        creator.clone(),
+        String::from_str(&env, "Cleanup Failed"),
+        String::from_str(&env, "Failed campaign refund"),
+        10_000,
+        30,
+        Category::Learner,
+        false,
+        0,
+        0i128,
+    ));
+    client.verify_campaign(&id);
+    client.contribute(&id, &contributor1, &500);
+
+    env.ledger().with_mut(|li| {
+        li.timestamp += 31 * 86400;
+    });
+
+    client.claim_refund(&id, &contributor1);
+
+    assert!(
+        !has_persistent_key(
+            &env,
+            &client,
+            DataKey::Contribution(id, contributor1.clone())
+        ),
+        "Contribution key must be removed after refund on failed campaign"
+    );
+}
+
 // Issue #341: claim_revenue is gated on funds_withdrawn. The prior "claim
 // then cancel then refund" flow this test exercised is now structurally
 // impossible (cancel is blocked once funds are withdrawn). Covered by
 // test::test_claim_revenue_blocked_before_funds_withdrawn.
+
+#[test]
+fn test_claim_revenue_amount_raised_zero_guard() {
+    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
+
+    token_admin.mint(&contributor1, &5_000);
+
+    let campaign_id = client.create_campaign(&make_params(
+        creator.clone(),
+        String::from_str(&env, "Zero Raised Guard"),
+        String::from_str(&env, "Directly test AmountRaisedIsZero guard"),
+        1_000,
+        30,
+        Category::EducationalStartup,
+        true,
+        2000,
+        0i128,
+    ));
+
+    client.verify_campaign(&campaign_id);
+    client.contribute(&campaign_id, &contributor1, &500);
+
+    // Artificially zero out amount_raised and effective_amount_raised while keeping
+    // the contribution in storage. Also force funds_withdrawn=true so we bypass the
+    // funds_withdrawn guard and exercise the AmountRaisedIsZero guard specifically.
+    env.as_contract(&client.address, || {
+        let mut campaign = storage::get_campaign(&env, campaign_id).unwrap();
+        campaign.amount_raised = 0;
+        campaign.effective_amount_raised = 0;
+        campaign.funds_withdrawn = true;
+        storage::set_campaign(&env, campaign_id, &campaign);
+    });
+
+    let res = client.try_claim_revenue(&campaign_id, &contributor1);
+    assert_eq!(res.unwrap_err().unwrap(), Error::AmountRaisedIsZero);
+}
+
+#[test]
+fn test_claim_refund_clears_lifetime_contribution() {
+    let (env, _admin, creator, contributor1, _, _token, token_admin, client) = setup_env();
+
+    token_admin.mint(&contributor1, &5000);
+
+    let campaign_id = client.create_campaign(&make_params(
+        creator.clone(),
+        String::from_str(&env, "LT Cleanup"),
+        String::from_str(&env, "Test lifetime cleanup"),
+        2000,
+        1,
+        Category::Learner,
+        false,
+        0,
+        1_000i128,
+    ));
+    let _ = client.try_verify_campaign(&campaign_id);
+    client.contribute(&campaign_id, &contributor1, &900);
+
+    // Cancel and refund
+    client.cancel_campaign(&campaign_id);
+    client.claim_refund(&campaign_id, &contributor1);
+
+    // LifetimeContribution should be cleared after full refund
+    assert_eq!(
+        client.get_lifetime_contribution(&campaign_id, &contributor1),
+        0,
+        "LifetimeContribution should be 0 after full refund"
+    );
+}
